% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedSim_Functions.R
\name{add_mate}
\alias{add_mate}
\title{Create a new mating partner}
\usage{
add_mate(partner_info, last_id)
}
\arguments{
\item{partner_info}{All ped file information for the individual requiring a
new mate, must contain all fields generated by \link{create_pedFile}.}

\item{last_id}{numeric.  The last id used in the pedigree.}
}
\value{
new_mate_info the ped file information for the new mate

last_id the updated value of last_id after mate has been added
}
\description{
Create a new mating partner
}

