% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedSim_Functions.R
\name{add_offspring}
\alias{add_offspring}
\title{Create a new offspring.}
\usage{
add_offspring(dad_info, mom_info, byear, last_id)
}
\arguments{
\item{dad_info}{All ped file information of offspring's father, must contain all fields generated by \code{\link{create_pedFile}}.}

\item{mom_info}{All ped file information of offspring's mother, must contain all fields generated by \code{\link{create_pedFile}}.}

\item{byear}{The birth year of offspring.}

\item{last_id}{numeric.  The last id used in the pedigree.}
}
\value{
new_child_info the ped file information for the new child

last_id the updated value of last_id after offspring has been added
}
\description{
Create a new offspring for pedigree.
}

