% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManipulatePed_Functions.R
\name{affInfo.ped}
\alias{affInfo.ped}
\title{Obtain specialized pedigree information}
\usage{
affInfo.ped(ped_file)
}
\arguments{
\item{ped_file}{Data.frame. A pedigree generated by \code{sim_ped} or \code{sim_RVped}.}
}
\value{
A list containing the following:

\item{\code{affVars}}{A data.frame containing information for affected relatives only.}

\item{\code{affKin}}{A list containing matrices of the pariwise kinship  coefficients for the affected relatives for each family in \code{ped_file.}}
}
\description{
Obtain specialized pedigree information
}
\details{
Users who wish to use \code{affInfo.ped} for pedigrees not simulated with the \code{SImRVPedigree} package must ensure that the pedigree, \code{ped_file}, supplied to \code{affInfo.ped} contains the following variables for each pedigree member:
\enumerate{
\item \code{ID}: an identification number.
\item \code{dadID}: identification number of father.
\item \code{momID}: identification number of mother.
\item \code{sex}: sex identification; if male \code{sex = 0}, if female \code{sex = 1}.
\item \code{affected}: affection status, if affected by disease \code{affected  = 1}, otherwise, \code{affected = 0}.}
Optionally, \code{ped_file} may contain any of the additional variables contained in pedigrees simulated by \code{\link{sim_RVped}}.
}
\examples{
#Read in example pedigree to trim
data(AgeSpecific_Hazards)

haz_obj <- hazard(hazardDF = AgeSpecific_Hazards)

#Simulate a pedigree ascertained for multiple affecteds
set.seed(8237)
RVped2015 <- sim_RVped(hazard_rates = haz_obj,
                       num_affected = 2,
                       ascertain_span = c(1900, 2015),
                       GRR = 30, carrier_prob = 0.002,
                       RVfounder = TRUE,
                       stop_year = 2015,
                       recall_probs = c(1),
                       founder_byears = c(1900, 1905),
                       FamID = 1)[[2]]
summary(RVped2015)
Ainfo <- affInfo.ped(RVped2015)

data(EgPeds)
egpeds <- ped(EgPeds)
Ainfo <- affInfo.ped(ped_file = egpeds)

#information for the affected relatives
Ainfo[[1]]

# matrix of pairwise kinship coefficents
# for the affected relatives in family 4
Ainfo[[2]][[4]]

}
\references{
Terry M Therneau and Jason Sinnwell (2015). **kinship2: Pedigree Functions.** *R package version 1.6.4.* https://CRAN.R-project.org/package=kinship2
}
