% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManipulatePed_Functions.R
\name{affKinM}
\alias{affKinM}
\title{Compute kinship matrix for the affected relatives}
\usage{
affKinM(ped_file)
}
\arguments{
\item{ped_file}{ped object}
}
\value{
\item{\code{affKin}}{The kinship matix for the affected relatives only.}
}
\description{
Compute kinship matrix for the affected relatives
}
\keyword{internal}
