% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHPP_functions.R
\name{approxCumHaz}
\alias{approxCumHaz}
\title{Determine cumulative risk at a given time}
\usage{
approxCumHaz(t, hazard, part)
}
\arguments{
\item{t}{A numeric constant. The age at which to approximate the cumulative hazard}

\item{hazard}{A numeric vector.  A vector of age-specific hazards.}

\item{part}{A numeric vector.  Partition of ages over which to apply the age-specific hazards.}
}
\value{
cum_hazard numeric. The approximate cumulative hazard at time t.
}
\description{
\code{approxCumHaz} returns the cumulative risk at time t, given a vector of age-specific hazards and a partition of ages over which to apply the age-specific hazards.
}
\examples{
haz_vec <- seq(1, 2.5, by = 0.1)
part_vec <- seq(0, 80, by = 5)
approxCumHaz(t = 0, hazard = haz_vec, part = part_vec)
approxCumHaz(t = 5, hazard = haz_vec, part = part_vec)
approxCumHaz(t = 17, hazard = haz_vec, part = part_vec)
approxCumHaz(t = 80, hazard = haz_vec, part = part_vec)
approxCumHaz(t = 100, hazard = haz_vec, part = part_vec)

}

