% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHPP_functions.R
\name{approxHazFun}
\alias{approxHazFun}
\title{create a cumulative hazard function that linearly interpolates the cumulative risk within a time interval}
\usage{
approxHazFun(hazard, part)
}
\arguments{
\item{hazard}{A numeric vector.  A vector of age-specific hazards.}

\item{part}{A numeric vector.  Partition of ages over which to apply the age-specific hazards.}
}
\description{
\code{approxHazFun} approximate a continuous cumulative hazard by linearly interpolating the cumulative risk within a time interval
}
\examples{
hfunct = approxHazFun(hazard = seq(1,2.5, by = 0.1),
                      part = seq(0, 80, by = 5))
hfunct(0)
hfunct(55)
hfunct(100)
}

