% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedCheck_functions.R
\name{ascertain_trimmedPed}
\alias{ascertain_trimmedPed}
\title{Check to see if a trimmed pedigree is ascertained.}
\usage{
ascertain_trimmedPed(ped_file, num_affected)
}
\arguments{
\item{ped_file}{Pedigree simulated by \code{sim_ped}.}

\item{num_affected}{Numeric.  The minimum number of affected individuals in the pedigree.}
}
\value{
Logical. If TRUE, pedigree is ascertained.
}
\description{
Check to see if a trimmed pedigree is ascertained.
}
\keyword{internal}
