% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVPedSim_Functions.R
\name{ascertainTrim_ped}
\alias{ascertainTrim_ped}
\title{Check to see if a trimmed pedigree is ascertained.}
\usage{
ascertainTrim_ped(ped_file, num_affected, first_diagnosis = NULL)
}
\arguments{
\item{ped_file}{Pedigree simulated by \code{sim_ped}.}

\item{num_affected}{Numeric.  The minimum number of affected individuals in the pedigree.}

\item{first_diagnosis}{Numeric. The first year that a reliable diagnosis could be made regarding disease-affection status.  By default, \code{first_diagnosis}\code{ = NULL} so that all diagnoses are considered valid. See details.}
}
\value{
Logical. If TRUE, pedigree is ascertained.
}
\description{
Check to see if a trimmed pedigree is ascertained.
}
\keyword{internal}
