% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{assign_gen}
\alias{assign_gen}
\title{Assign generation number based on oldest founder}
\usage{
assign_gen(x)
}
\arguments{
\item{x}{an object of class ped}
}
\value{
a list of generation numbers for pedigree members, in the order listed in \code{x}.
}
\description{
Assign generation number based on oldest founder
}
