% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManipulatePed_Functions.R
\name{assignGeneration}
\alias{assignGeneration}
\title{Assign minimum carrier generation to a simulated pedigree}
\usage{
assignGeneration(ped_file)
}
\arguments{
\item{ped_file}{data.frame. A ped file with the same format as one simulated with sim_RVpedigree}
}
\value{
reGen_ped A ped file with only affecteds and obligate carriers, with generation number assigned based on affected status.
}
\description{
\code{AssignGeneration} reassigns the generation number based on affected status and trims the pedigree down to affecteds and obligate carriers.
}
\details{
Add details concerning reassignment...
}
\examples{
part_vec <- seq(0, 100, by = 1)
unaffected_mort <- 0.00001 + pgamma(seq(0.16, 16, by = .16),
                                    shape = 9.5, scale = 1)/350
affected_mort <- c(0.55, 0.48, 0.37, 0.23, 0.15,
                   pgamma(seq(0.96, 16, by = .16),
                   shape = 4, scale = 1.5))/300
Dhaz_df  <- (as.data.frame(cbind(unaffected_mort, affected_mort)))
Ohaz_vec <- (dgamma(seq(0.1, 10, by = .1), shape = 8, scale = 0.75))/60

set.seed(1)
ex_RVped <- sim_RVpedigree(onset_hazard = Ohaz_vec,
                           death_hazard = Dhaz_df,
                           part = part_vec, RR = 5, FamID = 1,
                           founder_byears = c(1900, 1910),
                           ascertain_span = c(1900, 2015),
                           num_affected = 2)

plot_RVpedigree(ex_RVped[[1]])
plot_RVpedigree(ex_RVped[[2]])
}

