% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManipulatePed_Functions.R
\name{censor.ped}
\alias{censor.ped}
\title{Censor pedigree after a specified year}
\usage{
censor.ped(ped_file, censor_year)
}
\arguments{
\item{ped_file}{Data.frame. The pedigree to censor, see details.}

\item{censor_year}{Numeric. The censor year. If missing, when pedigree contains a proband, \code{censor_year} will assume the value of the proband's onset year. See details.}
}
\value{
\item{\code{censored_ped}}{The censored pedigree.}
}
\description{
\code{censor.ped} censors a pedigree of any information that occurs after a specified year.
}
\details{
Upon supplying a pedigree and a censor year the \code{censor.ped} function will remove all individuals born after \code{censor_year} and censor all disease onset and death events after the \code{censor_year}.

Users who wish to censor pedigrees which have not been simulated by \code{\link{sim_RVped}} or \code{\link{sim_ped}} must ensure that the pedigree, \code{ped_file}, supplied to \code{censor.ped} contains the following variables for each pedigree member:

\enumerate{
\item \code{ID}: an identification number.
\item \code{dadID}: identification number of father.
\item \code{momID}: identification number of mother.
\item \code{sex}: sex identification; if male \code{sex = 0}, if female \code{sex = 1}.
\item \code{affected}: affection status, if affected by disease \code{affected  = 1}, otherwise, \code{affected = 0}.
\item \code{birthYr}: the individual's birth year.
\item \code{onsetYr}: the individual's disease onset year, when applicable.
\item \code{deathYr}: the individual's death year, when applicable.
\item \code{proband}: (Optional) Proband identification variable, \code{proband = 1} if the individual is the proband, and 0 otherwise.
}

If an individual has not experienced disease onset and/or death, then \code{onsetYr = NA} and/or \code{deathYr = NA}.

If \code{censor_year} is missing, when the pedigree contains a proband, \code{censor_year} is set, internally, to the year that the proband experienced disease onset. However, if \code{ped_file} does not contain the proband identification variable the user must supply a value for \code{censor_year}.

For a detailed example please refer to vignette.
}
\examples{
#Read in example pedigree to trim
data(AgeSpecific_Hazards)

haz_obj <- hazard(hazardDF = AgeSpecific_Hazards)

#Simulate a pedigree ascertained for multiple affecteds
set.seed(3)
RVped2015 <- sim_RVped(hazard_rates = haz_obj,
                       num_affected = 2,
                       ascertain_span = c(1900, 2015),
                       GRR = 30, carrier_prob = 0.002,
                       RVfounder = TRUE,
                       stop_year = 2015,
                       recall_probs = c(1),
                       founder_byears = c(1900, 1905),
                       FamID = 1)[[2]]

#Censor ex_RVped after 1960
RVped1960 <- censor.ped(ped_file = RVped2015, censor_year = 1960)

#To plot, first convert to kinship2 pedigree class
pedigree2015 <- ped2pedigree(RVped2015)
pedigree1960 <- ped2pedigree(RVped1960)

#To plot pedigrees, use the kinship2 package
library(kinship2)

par(mfrow = c(1, 2))
plot(pedigree2015)
pedigree.legend(pedigree2015, location = "bottomleft",  radius = 0.25)
mtext(side = 3, "Reference Year: 2015")

plot(pedigree1960)
mtext(side = 3, "Reference Year: 1960")
par(mfrow = c(1, 1))


}
