% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_and_warning_functions.R
\name{check_hazpart}
\alias{check_hazpart}
\title{Check that age-specific hazard and partition of ages are correctly specified.}
\usage{
check_hazpart(hazard, part)
}
\arguments{
\item{hazard}{A numeric vector.  A vector of age-specific hazards.}

\item{part}{A numeric vector.  Partition of ages over which to apply the
age-specific hazards.}
}
\description{
Check that age-specific hazard and partition of ages are correctly specified.
}
\examples{
#the length of part should be equal to the length of hazard + 1
check_hazpart(hazard = seq(0, 1, by = 0.05), part = seq(0, 21, by = 1))
\dontrun{check_hazpart(hazard = seq(0, 1, by = 0.05),
                       part = seq(0, 20, by = 1))}
\dontrun{check_hazpart(hazard = c(1, NA, 2, 4),
                       part = c(0, 10, 20, 30, 40))}

}

