% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_and_warning_functions.R
\name{check_part}
\alias{check_part}
\title{define a function that issues a warning if age-specific hazards are notsupplied for an appropriate span of ages}
\usage{
check_part(part)
}
\arguments{
\item{part}{A numeric vector. Partiton over which to apply hazards}
}
\description{
\code{check_part} Check to see if the partition vector meets some basic properties
}
\examples{
check_part(part = c("a"))
check_part(part = c(-1, 0, 10))
}

