% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_and_warning_functions.R
\name{check_rprobs}
\alias{check_rprobs}
\title{Check to see if recall probabilies are correctly specified}
\usage{
check_rprobs(recall_probs)
}
\arguments{
\item{recall_probs}{Numeric.  Proband recall probabilities.  See details.}
}
\description{
Check to see if recall probabilies are correctly specified
}
\examples{
\dontrun{check_rprobs(c(2, 0.3, 0.4))
         check_rprobs(NA)
         check_rprobs("a")}
check_rprobs(recall_probs = c(1))
}

