% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_and_warning_functions.R
\name{check_spans}
\alias{check_spans}
\title{Check to see if time spans are appropriately specified}
\usage{
check_spans(span, part)
}
\arguments{
\item{span}{numeric.  A span of years or ages.}
}
\description{
Check to see if time spans are appropriately specified
}
\examples{
check_spans(c(1, 3, 6))
check_spans(c(1982, 1975))
check_spans(c(1975, 1982))
check_spans(c(1975, 1975))
check_spans(c(18))
}

