% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedSim_Functions.R
\name{create_mate}
\alias{create_mate}
\title{Create a new mate}
\usage{
create_mate(partner_info, last_id, GRR, prob_causalRV, single_founderEntry,
  RV_founder)
}
\arguments{
\item{partner_info}{Data.frame with 1 row. All pedigree information for the individual requiring a new mate, must contain all fields generated by \code{\link{create_pedFile}}.}

\item{last_id}{Numeric.  The last id used in the pedigree.}

\item{GRR}{Numeric. The relative-risk of disease for individuals who inherit the causal variant.}

\item{prob_causalRV}{Numeric.  The population allele frequency of the genetic variant with relative risk \code{GRR}. See details.}

\item{single_founderEntry}{Logical. When \code{single_founderEntry = TRUE}, only 1 founder can introduce a causal variant to the pedigee; when \code{single_founderEntry = FALSE} multiple founders may introduce the causal variant.  By default, \code{single_founderEntry = TRUE}.}

\item{RV_founder}{Logical. If \code{RV_founder = TRUE}, then at least 1 founder has introduced the causal variant to the pedigree. If \code{RV_founder = FALSE} then the variant has not yet been introduced to the pedigree.}
}
\value{
\code{new_mate_info} the pedigree information for the new mate.

\code{last_id} the updated value of last_id after adding the new mate.

\code{RV_founder} a true/false value indicating if at least 1 founder has introduced the RV.
}
\description{
Create new mating partner in pedigree.
}
\keyword{internal}
