% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedSim_Functions.R
\name{create_offspring}
\alias{create_offspring}
\title{Create a new offspring.}
\usage{
create_offspring(dad_info, mom_info, byear, last_id, RR)
}
\arguments{
\item{dad_info}{a data.frame with 1 row. All ped file information of offspring's father, must contain all fields generated by \code{\link{create_pedFile}}.}

\item{mom_info}{a data.frame with 1 row. All ped file information of offspring's mother, must contain all fields generated by \code{\link{create_pedFile}}.}

\item{byear}{The birth year of offspring.}

\item{last_id}{numeric.  The last id used in the pedigree.}

\item{RR}{Numeric. The relative risk of disease onset for individuals who have inherited the rare variant.}
}
\value{
new_child_info the ped file information for the new child

last_id the updated value of last_id after adding new offspring
}
\description{
Create new offspring in pedigree.
}

