% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EgPeds.R, R/ExamplePedigrees.R
\docType{data}
\name{EgPeds}
\alias{EgPeds}
\title{Example Pedigrees}
\format{A data frame with 50 rows and 14 variables:
\describe{
\item{FamID}{Family identification number}
\item{ID}{Individual identification number}
\item{gender}{Gender identification variable: \code{gender = 0} for males, and \code{gender = 1} females. }
\item{dad_id}{Identification number of father}
\item{mom_id}{Identification number of mother}
\item{affected}{Affection status: \code{affected = 1} if individual has developed lymphoid cancer, and 0 otherwise.}
\item{DA1}{Paternally inherited allele at the assumed disease locus: \code{DA1 = 1} if rare variant is present, and 0 otherwise.}
\item{DA2}{Maternally inherited allele at the assumed disease locus: \code{DA2 = 1} if rare variant is present, and 0 otherwise.}
\item{birth_year}{Subject's year of birth}
\item{onset_year}{Subject's year of disease onset, when applicable.}
\item{death_year}{Subject's year of death, when applicable.}
\item{RR}{Subject's relative risk of disease}
\item{available}{Availability status, \code{available  = 0} if the individual is unavailable, and 1 otherwise}
\item{Gen}{The subject's generation number relative to the founder who introduced the rare variant.  That is, the founder who introduced the rare variant will have \code{Gen = 1}, his or her offspring will have \code{Gen = 2}, etc.}
\item{proband}{Proband identification variable, \code{proband = 1} if the individual is the proband, and 0 otherwise.}
}}
\usage{
data(EgPeds)

data(EgPeds)
}
\description{
A dataset containing four example pedigrees.

A dataset containing four example pedigrees.
}
\keyword{datasets}

