% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeSim_Functions.R
\name{event_step}
\alias{event_step}
\title{Simulate the next life event for an individual}
\usage{
event_step(current_age, disease_status, RV_status, lambda_birth, onset_hazard,
  death_hazard, part, birth_range, RR)
}
\arguments{
\item{current_age}{The individuals current age.}

\item{disease_status}{The disease status, disease_status = 1 if individual
has experienced onset, and 0 otherwise.}

\item{RV_status}{The individuals rare variant status, RV_status  = 1 if
individual has inherited the rare variant, and 0 otherwise.}

\item{lambda_birth}{Numeric constant. The individuals birth rate.}

\item{onset_hazard}{Numeric vector. The population age-specific onset hazard.}

\item{death_hazard}{data.frame. Column 1 should specify the age specific
mortality rates in the unaffected population, while column 2 should provide
the age specific morality rates in the affected population.}

\item{part}{A numeric vector.  Partition of ages over which to apply the
age-specific hazards}

\item{birth_range}{A numeric vector of length 2.  The minimum and maximum
allowable birth ages in simulation.}

\item{RR}{A numeric constant. The rare variant relative risk of developing
disease.}
}
\description{
\code{event_step} Randomly simulates the next life event for an individual by
generating the waiting times to reproduction, onset, and death given the
individuals current age and then choosing the event with the shortest waiting
time as the winner.
}

