% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event_Methods.R
\name{events}
\alias{events}
\title{Constructor function for an object of class events}
\usage{
events(life_events)
}
\arguments{
\item{life_events}{The list of items returned by the \code{sim_life} function.}
}
\value{
an object of class \code{events}.
}
\description{
Constructor function for an object of class events
}
\keyword{internal}
