% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{find_available_parent}
\alias{find_available_parent}
\title{Find available parent}
\usage{
find_available_parent(ped, ID)
}
\arguments{
\item{ped}{the ped object}

\item{ID}{the Id of the person whose availble parent we would like to find}
}
\value{
the ID of the availble parent
}
\description{
Find available parent
}
