% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{find_missing_parent}
\alias{find_missing_parent}
\title{Find parents that have been removed but are required for plotting.}
\usage{
find_missing_parent(x, dad = TRUE)
}
\arguments{
\item{x}{An object of class ped.}

\item{dad}{Logical. If TRUE looking for missing dad, if FALSE looking for missing mom.}
}
\value{
The IDs of the missing parents.
}
\description{
Find parents that have been removed but are required for plotting.
}
\keyword{internal}
