% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{find_mrca}
\alias{find_mrca}
\title{Find the most recent common ancestor of two related individuals}
\usage{
find_mrca(ped, ID1, ID2)
}
\arguments{
\item{ped}{The ped object}

\item{ID1}{The ID of the first relative}

\item{ID2}{The ID of the second relative}
}
\value{
The ID of the common ancestor
}
\description{
Find the most recent common ancestor of two related individuals
}
