% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHPP_functions.R
\name{findWaitTime}
\alias{findWaitTime}
\title{Randomly generate the waiting time to next event, given the last event time}
\usage{
findWaitTime(u, last_event, hazard, part, scale = FALSE)
}
\arguments{
\item{u}{A numeric constant. argument of inverse CDF}

\item{last_event}{A numeric constant.  The age at last event.}

\item{hazard}{A numeric vector.  A vector of age-specific hazards.}

\item{part}{A numeric vector.  Partition of ages over which to apply the
age-specific hazards.}

\item{scale}{Logical. By default scale = FALSE.  NOTE: Choosing scale = TRUE ensures that W is a proper random variable, i.e. that this function is a proper CDF with upper limit 1.}
}
\value{
A numeric constant.  The waiting time to next event.
}
\description{
Randomly generate the waiting time to next event, given the last event time
}
\examples{
haz_vec <- c(seq(0, 0.5, by = 0.05), rev(seq(0.46, 0.5, by = 0.01)))
part_vec <- seq(0, 80, by = 5)

set.seed(17)
u_val = runif(1)
findWaitTime(u = u_val, last_event = 2,
             hazard = haz_vec, part = part_vec)

findWaitTime(u = u_val, last_event = 2,
             hazard = haz_vec, part = part_vec,
             scale = TRUE)

findWaitTime(u = u_val, last_event = 10,
             hazard = haz_vec, part = part_vec,
             scale = TRUE)

}

