% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManipulatePed_Functions.R
\name{get_affectedInfo}
\alias{get_affectedInfo}
\title{Gather information for the affected relatives}
\usage{
get_affectedInfo(ped_file)
}
\arguments{
\item{ped_file}{ped object}
}
\value{
\item{\code{affVars}}{Information for the affected relatives.}
}
\description{
Gather information for the affected relatives
}
\keyword{internal}
