% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedSim_Functions.R
\name{get_eventYear}
\alias{get_eventYear}
\title{Convert event waiting times to year of event}
\usage{
get_eventYear(Rlife_events, byear)
}
\arguments{
\item{Rlife_events}{Named matrix of all life events returned from
\code{\link{get_lifeEvents}}}

\item{byear}{The birth year}
}
\value{
Named matrix. The years when simulated life events occur,
named according to which life event has occurred.
}
\description{
Convert event waiting times to year of event
}

