% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManipulatePed_Functions.R
\name{get_famInfo}
\alias{get_famInfo}
\title{Obtain summary information}
\usage{
get_famInfo(ped_file, s_ID = NULL)
}
\arguments{
\item{ped_file}{A ped object}

\item{s_ID}{A vector of subtype IDs. By default, \code{s_ID = NULL}, which implies there are no subtypes.}
}
\value{
A data frame containing the relevant summary information
}
\description{
Obtain summary information
}
\keyword{internal}
