% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{get_gen_labs}
\alias{get_gen_labs}
\title{Get generation lables}
\usage{
get_gen_labs(x, nlevel)
}
\arguments{
\item{x}{An object of class ped.}

\item{nlevel}{The total number of levels in the pedigree.}
}
\value{
A character list of either blanks or roman numerals which wil be plotted along side the pedigree.
}
\description{
Get generation lables
}
