% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_Functions.R
\name{get_ind_data}
\alias{get_ind_data}
\title{Get individual data for sim_pop function}
\usage{
get_ind_data(hazard_rates, GRR, YOB, stop_year, RVstat, carrier_prob)
}
\arguments{
\item{hazard_rates}{An object of class \code{hazard}, created by \code{\link{hazard}}.}

\item{GRR}{Numeric. The genetic relative-risk of disease, i.e. the relative-risk of disease for individuals who carry at least one copy of the causal variant.}

\item{YOB}{Numeric. The year of birth for all individuals in the sample.}

\item{stop_year}{Numeric. The last year of the study, i.e. the last year we can observe data.}

\item{RVstat}{The individual's cRV status}

\item{carrier_prob}{Numeric.  The carrier probability for all causal variants with relative-risk of disease \code{GRR}.  By default, \code{carrier_prob}\code{ = 0.002}}
}
\value{
A data frame of survival data
}
\description{
This is an internal function
}
\keyword{internal}
