% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeSim_Functions.R
\name{get_lifeEvents}
\alias{get_lifeEvents}
\title{Simulate all life events.}
\usage{
get_lifeEvents(onset_hazard, death_hazard, part, birth_range, NB_params, RR,
  YOB)
}
\arguments{
\item{onset_hazard}{Numeric. The population age-specific hazard rate for disease onset.}

\item{death_hazard}{Data.frame. Column 1 should specify the unaffected age-specific hazard rate for death, and column 2 should specify the affected age-specific hazard rate for death.}

\item{part}{Numeric. The partition of ages over which to apply the age-specific hazard rates in \code{onset_hazard} and \code{death_hazard}.}

\item{birth_range}{Numeric list of length 2. By default, \code{birth_range = c(18, 45)}. The minimum and maximum allowable ages that individuals will be able to reproduce.}

\item{NB_params}{Numeric list of length 2. By default, \code{NB_params = c(2, 4/7)}. The size and probability parameters of the negative binomial distribution used to model the number of children per household.}

\item{RR}{Numeric. The relative risk of disease onset for individuals who have inherited the rare variant.}

\item{YOB}{Numeric. The indivdiual's year of birth.}
}
\value{
Named matrix. The waiting times between all life events simulated for an individual, named according to which life event has occurred.
}
\description{
\code{get_lifeEvents} simulates all life events for an individual starting at age
0 and ending with death by applying the \link{get_nextEvent} function until death
occurs.
}
\examples{
part_vec <- seq(0, 100, by = 1)
unaffected_mort <- 0.00001 + pgamma(seq(0.16, 16, by = .16),
                                    shape = 9.5, scale = 1)
affected_mort <- c(0.55, 0.48, 0.37, 0.23, 0.15,
                   pgamma(seq(0.96, 16, by = .16), shape = 4, scale = 1.5))
Dhaz_df  <- (as.data.frame(cbind(unaffected_mort, affected_mort)))
Ohaz_vec <- (dgamma(seq(0.1, 10, by = .1), shape = 8, scale = 0.75))
set.seed(17)
get_lifeEvents(onset_hazard = Ohaz_vec,
               death_hazard = Dhaz_df,
               part = part_vec,
               birth_range = c(17,45),
               NB_params = c(2, 4/7), RR = 1,
               YOB = 1900)

set.seed(17)
get_lifeEvents(onset_hazard = Ohaz_vec,
               death_hazard = Dhaz_df,
               part = part_vec,
               birth_range = c(17,45),
               NB_params = c(2, 4/7), RR = 15,
               YOB = 1900)

}

