% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHPP_functions.R
\name{get_WaitTime}
\alias{get_WaitTime}
\title{Obtain the waiting time to next event associated with a nonhomogeneous Poisson process.}
\usage{
get_WaitTime(p, last_event, hazard, part, scale = FALSE)
}
\arguments{
\item{p}{a probability.  Argument of the quantile function of the waiting time associated with a non-homogeneous poisson process with rate \code{hazard}.}

\item{last_event}{a positive number.  The time at last event.}

\item{hazard}{A vector of positive numbers.  The time-specific hazard rate.}

\item{part}{A vector of positive numbers.  Time partition over which to apply the time-specific hazard rate.}

\item{scale}{logical. By default scale = FALSE.  Specifying scale = TRUE ensures that this function is the quantile function for a proper random variable.}
}
\value{
The waiting time to next event, units same as those in \code{part}.
}
\description{
\code{get_WaitTime} approximates the result of the inverse cumulative distribution function of the waiting time to next event associated with a non-homogeneous Poisson process conditioned of the time of the last event.
}
\details{
\code{get_WaitTime} obtains the waiting time to next event associated with a non-homogeneous Poisson process.  The units of the simulated waiting time are the units specified in \code{part}, i.e. if \code{part} is specified in years, the simulated waiting time is in years.
}
\examples{
haz_vec <- c(seq(0, 0.5, by = 0.05), rev(seq(0.46, 0.5, by = 0.01)))
part_vec <- seq(0, 80, by = 5)

get_WaitTime(p = 0.05, last_event = 0,
             hazard = haz_vec, part = part_vec,
             scale = FALSE)

get_WaitTime(p = 0.05, last_event = 0,
             hazard = haz_vec, part = part_vec,
             scale = TRUE)
}

