% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hazard_object_functions.R
\name{hazard}
\alias{hazard}
\title{Constructor function for an object of class \code{hazard}}
\usage{
hazard(hazardDF, partition)
}
\arguments{
\item{hazardDF}{Data.frame. Column 1 should specify the age-specific hazard rate for individuals with no genetic predisposition to the disease, column 2 should specify the age-specific hazard rate for death in the unaffected population, and column 3 should specify the age-specific hazard rate for death in the affected population.}

\item{partition}{Numeric vector. Optional. The partition of ages over which to apply the age-specific hazard rates in \code{hazardDF}. If missing, it is assumed that \code{partition} starts at 0 and increases in yearly increments.}
}
\value{
an object of class \code{hazard}.
}
\description{
Constructor function for an object of class \code{hazard}
}
\examples{
my_HR <- hazard(partition = seq(0, 100, by = 1),
                hazardDF = AgeSpecific_Hazards)
class(my_HR)

}
