% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event_Methods.R
\name{is.events}
\alias{is.events}
\title{Check to see if object is of class ped}
\usage{
is.events(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
Logical. Indicates if \code{x} is of class \code{ped}.
}
\description{
Check to see if object is of class ped
}
\keyword{internal}
