% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hazard_Methods.R
\name{is.hazard}
\alias{is.hazard}
\title{Check to see if object is of class hazard}
\usage{
is.hazard(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
Logical. Indicates if \code{x} is of class \code{hazard}.
}
\description{
Check to see if object is of class hazard
}
\keyword{internal}
