% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedCheck_functions.R
\name{is_reAscertained}
\alias{is_reAscertained}
\title{Determine if a previously ascertained pedigree meets new criteria.}
\usage{
is_reAscertained(asc_ped, num_affected, ref_year)
}
\arguments{
\item{asc_ped}{Data.frame.  The previously ascertained pedigree. Optionally, asc_ped may contain ped_file information for the affected relatives only.}

\item{num_affected}{Numeric.  The minimum number of affected individuals in the pedigree.}

\item{ref_year}{A numeric constant.  The reference year used to determine current age for pedigree members.}
}
\value{
Logical. If TRUE, pedigree is re-ascertained.
}
\description{
Determine if a previously ascertained pedigree meets new criteria.
}
\examples{
#Read in age-specific hazards
data(AgeSpecific_Hazards)

my_HR <- hazard(hazardDF = AgeSpecific_Hazards)

#Simulate pedigree ascertained for multiple affected individuals
set.seed(8008135)
ex_ped <- sim_ped(hazard_rates = my_HR,
                  GRR = 50, carrier_prob = 0.002,
                  RVfounder = TRUE,
                  FamID = 1,
                  founder_byears = c(1900, 1910))

head(ex_ped, n = 3)
Aped <- is_ascertained(ped_file = ex_ped,
                       num_affected = 2,
                       ascertain_span = c(2000, 2015),
                       recall_probs = c(1, 1, 0.5, 0.25))

Aped[[1]]

Aped[[2]][(Aped[[2]]$affected == 1 & Aped[[2]]$available == 1), ]

is_reAscertained(asc_ped = Aped[[2]],
                 num_affected = 3,
                 ref_year = 2012)

is_reAscertained(asc_ped = Aped[[2]],
                 num_affected = 3,
                 ref_year = 2016)

}
