% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedSim_Functions.R
\name{nfam_step}
\alias{nfam_step}
\title{Simulate all life events for an individual and create a nuclear family when
appropriate.}
\usage{
nfam_step(found_info, stop_year, last_id, onset_hazard, death_hazard, part,
  birth_range, NB_params, RR)
}
\arguments{
\item{found_info}{All ped file information for the founder, must contain all fields generated by \link{create_pedFile}.}

\item{stop_year}{The last year of study.  Note: any information simulated after stop.year is censored and will not be recorded in the ped file.}

\item{last_id}{numeric.  The last id used in the pedigree.}

\item{onset_hazard}{Numeric vector. The population age-specific onset hazard.}

\item{death_hazard}{data.frame. Column 1 should specify the age specific
mortality rates in the unaffected population, while column 2 should provide
the age specific morality rates in the affected population.}

\item{part}{A numeric vector.  Partition of ages over which to apply the
age-specific hazards}

\item{birth_range}{A numeric vector of length 2.  The minimum and maximum
allowable birth ages in simulation.}

\item{NB_params}{A numeric vector of length 2. The size and probabiliy
parameters of the negative binomial distribution that describes the number of
children per household in the population.}

\item{RR}{A numeric constant. The rare variant relative risk of developing
disease.}
}
\value{
A ped file with all life events of founding individual updated, and
with additional information for mate and offspring, when offspring are generated. #'
}
\description{
Simulate all life events for an individual and create a nuclear family when
appropriate.
}

