% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ped_object_functions.R
\name{ped}
\alias{ped}
\title{Constructor function for an object of class ped}
\usage{
ped(ped_file)
}
\arguments{
\item{ped_file}{Data.frame. A pedigree, see details.}
}
\value{
an object of class \code{ped}.
}
\description{
Constructor function for an object of class ped
}
