% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManipulatePed_Functions.R
\name{pedigree_info}
\alias{pedigree_info}
\title{Return useful pedigree information}
\usage{
pedigree_info(ped_file)
}
\arguments{
\item{ped_file}{Data.frame. A pedigree generated by \code{sim_ped}.}
}
\value{
A list containing the following:

\code{link_format} data.frame. The pedigree in linkage format; i.e. containing only the fields: FamID, ID, dadID, momID, affected.

\code{affected_info} data.frame.  Relevant information for the affected realtives only.

\code{kin_ped} An object of class pedigree (see kinship2 package).  A pedigree object which can be plotted using R's plot function.  See example.

\code{aff_kinMat} The kinship matrix for the affected relatives in the pedigree (see kinship2 package).
}
\description{
Return useful pedigree information
}
\examples{
data(EgPeds)

Fam1 <- pedigree_summary(EgPeds[EgPeds$FamID == 1, ])
summary(Fam1)

head(Fam1$link_format)
Fam1$affected_info
Fam1$aff_kinMat

library(kinship2)
plot(Fam1$kin_ped)
pedigree.legend(Fam1$kin_ped, location = "topleft",  radius = 0.25)

}
\references{
Terry M Therneau and Jason Sinnwell (2015). **kinship2: Pedigree Functions.** *R package version 1.6.4.* https://CRAN.R-project.org/package=kinship2
}
