% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ped2Pedigree_functions.R
\name{pedigreeLabels}
\alias{pedigreeLabels}
\title{Create pedigree labels}
\usage{
pedigreeLabels(x, ref_year)
}
\arguments{
\item{x}{An object of class ped.}

\item{ref_year}{A numeric constant.  The reference year used to determine current age for pedigree members.}
}
\value{
A list of labels that can be used with kinship2's plot function. See \code{\link{plot.pedigree}} for details.
}
\description{
Create pedigree labels
}
\references{
Terry M Therneau and Jason Sinnwell (2015). \strong{kinship2: Pedigree Functions.} \emph{R package version 1.6.4.} https://CRAN.R-project.org/package=kinship2
}
\keyword{internal}
