% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Functions.R
\name{plot_RVpedigree}
\alias{plot_RVpedigree}
\title{Plot simulated pedigree}
\usage{
plot_RVpedigree(ped_file)
}
\arguments{
\item{ped_file}{A single pedigree simulated with either \code{sim_RVpedigree} or \code{ped_step}}
}
\value{
Values returned from \link[kinship2]{plot.pedigree}
}
\description{
Plot ped files generated using \code{SimRVPedigree} using the pedigree plotting tools provided by the \link{kinship2} package.
}
\examples{
part_vec <- seq(0, 100, by = 1)
unaffected_mort <- 0.00001 + pgamma(seq(0.16, 16, by = .16),
                                    shape = 9.5, scale = 1)/350
affected_mort <- c(0.55, 0.48, 0.37, 0.23, 0.15,
                   pgamma(seq(0.96, 16, by = .16), shape = 4, scale = 1.5))/300
Dhaz_df  <- (as.data.frame(cbind(unaffected_mort, affected_mort)))
Ohaz_vec <- (dgamma(seq(0.1, 10, by = .1), shape = 8, scale = 0.75))
set.seed(22)
ex_ped <- ped_step(onset_hazard = Ohaz_vec, death_hazard = Dhaz_df,
                   part = part_vec, RR = 5, founder_byears = c(1900, 1910))
ex_ped$FamID = 1
plot_RVpedigree(ex_ped)
}

