% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hazard_object_functions.R
\name{print.hazard}
\alias{print.hazard}
\title{Prints hazard object.}
\usage{
\method{print}{hazard}(x)
}
\arguments{
\item{x}{hazard object}
}
\description{
Prints \code{\link{hazard}} object.
}
