% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManipulatePed_Functions.R
\name{reassignGen.ped}
\alias{reassignGen.ped}
\title{Reassign generation number based on affected status}
\usage{
reassignGen.ped(ped_file)
}
\arguments{
\item{ped_file}{An object of class \code{ped}. A pedigree generated by \code{simPed} or \code{simRVped}, or an object created by the function \code{\link{new.ped}}.  See details.}
}
\value{
\item{\code{reGen_ped} }{A pedigree containing only affected members, obligate carriers, and founders with generation number based on the most recent common ancestor of affected members as, described in details.}
}
\description{
The \code{reassignGen.ped} function assigns generation numbers among affected family members so that generation 1 represents the most recent generation that a putative disease variant shared identical by descent (IBD), as defined in Thompson (2000), by affected members could have been introduced into the pedigree.
}
\details{
\emph{The \code{reassignGen.ped} function is primarily intended for pedigrees simulated by the \code{simRVPedigree} package.  Results will be inaccurate if imbreeding is present.}

The \code{reassignGen.ped} function accepts a pedigree simulated by \code{simRVped} and reassigns generation numbers among affected family members in the pedigree.  Specifically, given a pedigree this function reassigns the generation numbers of affected members so that generation 1 is assigned to the most recent common ancestor of all affected members.  We note that the individual in generation 1 could themselves be affected, i.e. an individual can be considered their own ancestor.

For example, consider a family with 2 affected members.  If the two affected members are a parent and his or her offspring, the affected parent would be assigned generation 1, and the affected child generation 2.  However, if the two affected members are a pair of siblings, each affected sibling would be assigned generation 2 since a common parent of the two affected siblings is assumed to be a carrier of a latent susceptibility variant.  Similarly, if the two affected members are a pair of cousins, each affected cousin is assigned generation 3, since a common grandparent of the two affected cousins is the most recent common ancestor from whom they could have inherited a shared variant associated with the disease.

Users who wish to assign generation number based on affection status in pedigrees that have not been simulated with the \code{SimRVpedigree} package must create a ped object using \code{new.ped}.  This \code{ped} object \emph{must} contain the following variables for each pedigree member:

\tabular{lll}{
\strong{name} \tab \strong{type} \tab \strong{description} \cr
\code{FamID} \tab numeric \tab family identification number \cr
\code{ID} \tab numeric \tab individual identification number \cr
\code{dadID} \tab numeric \tab identification number of father \cr
\code{momID} \tab numeric \tab identification number of mother \cr
\code{sex} \tab numeric \tab gender identification; if male \code{sex = 0}, if female \code{sex = 1} \cr
\code{affected  } \tab numeric \tab disease-affection status, \code{affected} = 1 if affected by disease, and 0 otherwise. \cr
\code{Gen} \tab numeric \tab the individual's generation number relative to the eldest founder. \cr
\tab \tab That is, for the eldest founder \code{Gen} = 1, for his or her offspring \code{Gen} = 2, etc. \cr
}
}
\examples{
# Read in example pedigrees
data(EgPeds)
class(EgPeds)

# Create ped object
Bpeds <- new.ped(EgPeds)
summary(Bpeds)

# Reassign generation numbers in the first four pedigrees in EgPeds
Apeds <- lapply(seq_len(4), function(x){
                 reassignGen.ped(Bpeds[Bpeds$FamID == x, ])})
Apeds <- do.call(rbind, Apeds)

# Create kinship2 pedigree objects so that we can plot pedigrees
kin2ped_before <- ped2pedigree(Bpeds)
kin2ped_after  <- ped2pedigree(Apeds)

# Compare pedigrees before and after reassigning
# generation number based on affected status
par(mfrow = c(1, 2))
for (k in 1:4) {
  ID1 = paste0("ID", sep = ":",
               Bpeds$ID[Bpeds$FamID == k],
               sep = "\\n Gen:", Bpeds$Gen[Bpeds$FamID == k])

  ID2 = paste0("ID", sep = ":",
               Apeds$ID[Apeds$FamID == k],
               sep = "\\n Gen:", Apeds$Gen[Apeds$FamID == k])

  plot(kin2ped_before[paste0(k)], id = ID1)
  mtext(paste0("Ped", k, ": before generation reassignment", sep = ""),
        side = 3)

  plot(kin2ped_after[paste0(k)], id = ID2)
  mtext(paste0("Ped", k, ": after generation reassignment", sep = ""),
        side = 3)
}
par(mfrow = c(1, 1))


}
\references{
OUR MANUSCRIPT

Thompson, E. (2000). \emph{Statistical Inference from Genetic Data on Pedigrees.} NSF-CBMS Regional Conference Series in Probability and Statistics, 6, I-169. Retrieved from http://www.jstor.org.proxy.lib.sfu.ca/stable/4153187
}
\seealso{
\code{\link{new.ped}}
}

