% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManipulatePed_Functions.R
\name{reduce_to_affected}
\alias{reduce_to_affected}
\title{Reduce pedigree to contain disease-affected relatives}
\usage{
reduce_to_affected(ped_file)
}
\arguments{
\item{ped_file}{A ped object. See \code{\link{new.ped}} for details.}
}
\value{
\code{red_ped} A pedigree containing only affected members, obligate carriers, and founders.
}
\description{
Reduce pedigree to contain only the disease-affected relatives and the individuals who
}
\examples{
library(SimRVPedigree)
#Read in example pedigrees and create ped object
data(EgPeds)
ex_peds <- new.ped(EgPeds)

#plot full pedigree
plot(ex_peds[which(ex_peds$FamID == 1), ])

#reduce to affected only pedigree
Apeds = reduce_to_affected(ex_peds[which(ex_peds$FamID == 1), ])
plot(Apeds)

}
