% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeSim_Functions.R
\name{sim_birthRate}
\alias{sim_birthRate}
\title{Simulate a random birth rate.}
\usage{
sim_birthRate(NB_size, NB_prob, min_birth_age, max_birth_age)
}
\arguments{
\item{NB_size, NB_prob}{size and probability parameters of negative binomial distribution.}

\item{min_birth_age, max_birth_age}{minimum and maximum allowable birth ages}
}
\value{
birth_rate numeric, the randomly generated birth rate
}
\description{
Simulate a random birth rate.
}
\examples{
set.seed(17)
sim_birthRate(NB_size = 2, NB_prob = 4/7,
              min_birth_age = 17, max_birth_age = 45)
set.seed(17)
sim_birthRate(NB_size = 2, NB_prob = 4/7,
              min_birth_age = 17, max_birth_age = 45)


}

