% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeSim_Functions.R
\name{sim_lifeEvents}
\alias{sim_lifeEvents}
\title{Simulate all life events}
\usage{
sim_lifeEvents(hazard_rates, GRR, allele_freq, RV_status, YOB, stop_year,
  birth_range, NB_params)
}
\arguments{
\item{hazard_rates}{An object of class \code{hazard}, created by \code{\link{new.hazard}}.}

\item{GRR}{Numeric. The genetic relative-risk of disease, i.e. the relative-risk of disease for individuals who carry at least one copy of the causal variant.}

\item{allele_freq}{Numeric.  The population allele frequency a genetic variant with relative risk of disease \code{GRR}.}

\item{RV_status}{Numeric. \code{RV_status = 1} if the individual is a carrier of a rare variant that increases disease suseptibility, and 0 otherwise.}

\item{YOB}{A positive number. The indivdiual's year of birth.}

\item{stop_year}{Numeric. The last year of study.  If missing, the current year is used.}

\item{birth_range}{Numeric vector of length 2. The minimum and maximum allowable ages, in years, between which individuals may reproduce.  By default, \code{birth_range = c(18, 45)}.}

\item{NB_params}{Numeric vector of length 2. The size and probability parameters of the negative binomial distribution used to model the number of children per household.  By default, \code{NB_params = c(2, 4/7)}, due to the investigation of Kojima and Kelleher (1962).}
}
\value{
A named matrix containing the years of an individual's simulated life events, named by event type, see details.
}
\description{
Primarily intended as an internal function, \code{sim_lifeEvents} simulates all life events for an individual starting at birth, age 0, and ending with death or the end of the study.
}
\details{
Starting at birth, age 0, \code{sim_lifeEvents} generates waiting times to reproduction, onset, and death. The event with the shortest waiting time is chosen as the next life event, and the individual's age is updated by the waiting time of the winning event.  Conditioned on the individual's new age, this process is applied recursively, until death or until the end of the study is reached.

We make the following assumptions regarding the simulation of waiting times:
 \enumerate{
 \item We assume that, given an individual's current age, their time to disease onset is the waiting time in a non-homogeneous Poisson process with an age-specific hazard rate that follows a proportional hazards model.  In this model, individuals who have NOT inherited the rare variant experience disease onset according to the baseline (or population) hazard rate of disease.  On the other hand, individuals who have inherited the rare variant are assumed to have an increased risk of disease onset relative to those who have inherited it.  The user is expected to supply the baseline hazard rate of disease, as well as the relative-risk of disease for genetic cases. Additionally, we impose the restriction that individuals may only experience disease onset once, and remain affected from that point on.

\item We assume that, given an individual's current age, their time to death is the waiting time in a non-homogeneous Poisson process with age-specific hazard rate determined by their affection status.  We assume that disease-affected individuals experience death according to the age-specific hazard rate for death in the \emph{affected} population.  On the other hand, we assume that \emph{unaffected} individuals experience death according to the age-specific hazard rate for death in the \emph{unaffected} population.  If the disease of interest is sufficiently rare, the user may choose to substitute the \emph{population} age-specific hazard rate for death for the aforementioned age-specific hazard rate for death in the \emph{unaffected} population.  The user is expected to supply age-specific hazard rates of death for both the \emph{affected} and \emph{unaffected} populations.

\item We assume that, given an individual's current age, their time to reproduction is the waiting time in a homogeneous Poisson process.  That is, we assume that individuals reproduce at uniform rate during their reproductive years.  For example, one's reproductive years may span from age 18 to age 45 years.  We do not allow for offspring to be produced outside of an individual's \code{birth_range}.

}

\code{sim_lifeEvents} will return a named matrix, which contains the years of the simulated life events, named by event type.  The possible event types are as follows:
\itemize{
 \item "Child" a reproductive event, i.e. creation of offspring
 \item "Onset" disease onset event,
 \item "Death" death event
}
}
\examples{
data(AgeSpecific_Hazards)

my_HR <- new.hazard(hazardDF = AgeSpecific_Hazards)

# The following commands simulate all life events for an individual, who
# has NOT inherited a causal variant, born in 1900.  From the output, this
# individual has 1 child in 1927, and then dies in 1987.
set.seed(7664)
sim_lifeEvents(hazard_rates = my_HR, GRR = 10,
               allele_freq = 0.02,
               RV_status = 0,
               birth_range = c(17,45),
               NB_params = c(2, 4/7),
               YOB = 1900, stop_year = 2000)

# Using the same random seed, notice how life events can vary for
# someone who has inherited the causal variant, which carries a
# relative-risk of 25. From the output, this individual also has
# 1 child in 1927, but then experiences disease onset in 1978,
# and dies in 1980.
set.seed(7664)
sim_lifeEvents(hazard_rates = my_HR, GRR = 10,
               allele_freq = 0.02,
               RV_status = 1,
               birth_range = c(17,45),
               NB_params = c(2, 4/7),
               YOB = 1900, stop_year = 2000)


}
\references{
OUR MANUSCRIPT

Ken-Ichi Kojima, Therese M. Kelleher. (1962), \emph{Survival of Mutant Genes}. The American Naturalist 96, 329-346.
}
