% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedSim_Functions.R
\name{sim_ped}
\alias{sim_ped}
\title{Simulate a pedigree}
\usage{
sim_ped(onset_hazard, death_hazard, part, RR, FamID, founder_byears,
  birth_range = c(18, 45), NB_params = c(2, 4/7), stop_year = 2015)
}
\arguments{
\item{onset_hazard}{Numeric vector. The population age-specific onset hazard.}

\item{death_hazard}{data.frame. Column 1 should specify the age specific
mortality rates in the unaffected population, while column 2 should provide
the age specific morality rates in the affected population.}

\item{part}{A numeric vector.  Partition of ages over which to apply the
age-specific hazards}

\item{RR}{A numeric constant. The relative risk of developing
disease for individuals who have inherited the rare variant.}

\item{FamID}{numeric. Family identification number.}

\item{founder_byears}{A numeric vector of length 2, the minimum and maximum founder birth years.  See details.}

\item{birth_range}{A numeric vector of length 2.  By default, \code{birth_range} = c(18, 45).  The minimum and maximum allowable birth ages in simulation.}

\item{NB_params}{A numeric vector of length 2. By default \code{NB_params} = c(2, 4/7). The size and probabiliy parameters of the negative binomial distribution that describes the number of children per household in the population.}

\item{stop_year}{The last year of study.  By default, \code{stop_year} = 2015. Note: any information simulated after stop.year is censored and will not be recorded in the ped file.}
}
\value{
The ped file for the simulated pedigree.
}
\description{
\code{sim_ped} randomly generates a pedigree. Note the distinction between \code{sim_ped} and \code{sim_RVpedigree}:  pedigrees simulated using \code{sim_ped} do not account for study design, pedigrees simulated using \code{sim_RVpedigrees} do account for study design.
}
\details{
\code{sim_ped} randomly generates a pedigree starting with the founder who introduced the rare variant.  We assume that the variant is rare enough that it has only been introduced to the pedigree by one of the two oldest founders.  The variant is transmitted to offspring according to Mendel's laws.  Life events for individuals who have inherited the rare variant (RV) are simulated such that their relative risk is \code{RR}, all others are simulated with a relative risk of 1.  All events which occur after the specified \code{stop_year} are censored.

The birth year for the founder with the rare variant is simulated uniformly between the years specified in \code{founder_byears}.
}
\examples{
part_vec <- seq(0, 100, by = 1)
unaffected_mort <- 0.00001 + pgamma(seq(0.16, 16, by = .16),
                                    shape = 9.5, scale = 1)/350
affected_mort <- c(0.55, 0.48, 0.37, 0.23, 0.15,
                   pgamma(seq(0.96, 16, by = .16), shape = 4, scale = 1.5))/300
Dhaz_df  <- (as.data.frame(cbind(unaffected_mort, affected_mort)))
Ohaz_vec <- (dgamma(seq(0.1, 10, by = .1), shape = 8, scale = 0.75))
set.seed(22)
ex_ped <- sim_ped(onset_hazard = Ohaz_vec, death_hazard = Dhaz_df,
                   part = part_vec, RR = 5, FamID = 1, founder_byears = c(1900, 1910))
}

