% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedSim_Functions.R
\name{sim_RVstudy}
\alias{sim_RVstudy}
\title{Simulate a dataset of n pedigrees ascertained for a rare disease}
\usage{
sim_RVstudy(npeds, onset_hazard, death_hazard, part, RR, founder_byears,
  ascertain_span, num_affected, recall_probs, first_FamID, birth_range = c(18,
  45), NB_params = c(2, 4/7), stop_year = 2015)
}
\arguments{
\item{npeds}{numeric. The number of pedigrees to simulate.}

\item{onset_hazard}{Numeric vector. The population age-specific onset hazard.}

\item{death_hazard}{data.frame. Column 1 should specify the age specific
mortality rates in the unaffected population, while column 2 should provide
the age specific morality rates in the affected population.}

\item{part}{A numeric vector.  Partition of ages over which to apply the
age-specific hazards}

\item{RR}{A numeric constant. The relative risk of developing
disease for individuals who have inherited the rare variant.}

\item{founder_byears}{A numeric vector of length 2, the minimum and maximum founder birth years.  See details.}

\item{ascertain_span}{numeric vector of length 2. The ascertainment period,
i.e. first and last year of study.}

\item{num_affected}{Numeric.  The desired number of affected.}

\item{recall_probs}{Numeric.  Proband recall probabilities.  See details.}

\item{birth_range}{A numeric vector of length 2.  By default, \code{birth_range} = c(18, 45).  The minimum and maximum allowable birth ages in simulation.}

\item{NB_params}{A numeric vector of length 2. By default \code{NB_params} = c(2, 4/7). The size and probabiliy parameters of the negative binomial distribution that describes the number of children per household in the population.}

\item{stop_year}{The last year of study.  By default, \code{stop_year} = 2015. Note: any information simulated after stop.year is censored and will not be recorded in the ped file.}
}
\value{
RVped_files A data.frame of n pedigrees.
}
\description{
\code{Sim_RVstudy} simulates a dataset of \emph{n} pedigrees that have been ascertained for a rare disease according to a particular study design.
}
\details{
For details on how individual pedigrees are simulated please refer to \link{sim_RVpedigree}.
}
\examples{
part_vec <- seq(0, 100, by = 1)
unaffected_mort <- 0.00001 + pgamma(seq(0.16, 16, by = .16),
                                    shape = 9.5, scale = 1)/350
affected_mort <- c(0.55, 0.48, 0.37, 0.23, 0.15,
                   pgamma(seq(0.96, 16, by = .16), shape = 4, scale = 1.5))/300
Dhaz_df  <- (as.data.frame(cbind(unaffected_mort, affected_mort)))
Ohaz_vec <- (dgamma(seq(0.1, 10, by = .1), shape = 8, scale = 0.75))
set.seed(22)
RVstudy = sim_RVstudy(npeds = 4, onset_hazard = Ohaz_vec,
                      death_hazard = Dhaz_df, part = part_vec, RR = 5,
                      founder_byears = c(1900, 1910),
                      ascertain_span = c(1900, 2015),
                      num_affected = 2 )

for(i in 1:4){
  plot_RVpedigree(RVstudy[which(RVstudy$FamID == i),])
  mtext(paste0("FamID = ", i))
}

}

