% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trimming_Functions.R
\name{trim_step}
\alias{trim_step}
\title{Choose proband and trim pedigree}
\usage{
trim_step(ped_file, ascertain_span, recall_probs)
}
\arguments{
\item{ped_file}{A data.frame. The ped file generated by \code{ped_step}.}

\item{ascertain_span}{numeric vector of length 2. The ascertainment period,
i.e. first and last year of study.}

\item{recall_probs}{numeric.  The proband recall probabilities.  See details.}
}
\value{
The trimmed ped file
}
\description{
\code{trim_step} chooses a proband from the pedigree and trims relatives based on proband recall probabilities.
}

