% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManipulatePed_Functions.R
\name{affInfo.ped}
\alias{affInfo.ped}
\title{Obtain information for disease-affected relatives}
\usage{
affInfo.ped(ped_file)
}
\arguments{
\item{ped_file}{An object of class \code{ped}. A pedigree generated by \code{simPed} or \code{simRVped}, or an object created by the function \code{\link{new.ped}}.  See details.}
}
\value{
A list containing the following:

\item{\code{affVars}}{A data.frame containing information for affected relatives in each family supplied to \code{affInfo.ped}.}

\item{\code{affKin}}{If multiple pedigrees supplied: a list containing matrices of the pariwise kinship coefficients for the disease-affected relatives for each family in \code{ped_file}.  If a single pedigree supplied: a matrix of the the pariwise kinship coefficients for the disease-affected relatives. See \code{\link{kinship}} for details.}
}
\description{
Obtain information for disease-affected relatives
}
\details{
Users who wish to use \code{affInfo.ped} for pedigrees not generated by \code{simPed} or \code{simRVped} must use \code{\link{new.ped}} to create an object of class \code{ped}.
}
\examples{
#Read in age-specific hazard data and create an object of class hazard.
data(AgeSpecific_Hazards)
haz_obj <- hazard(hazardDF = AgeSpecific_Hazards)

#Simulate a pedigree ascertained for multiple affecteds
set.seed(8237)
RVped2015 <- simRVped(hazard_rates = haz_obj,
                       num_affected = 2,
                       ascertain_span = c(1900, 2015),
                       GRR = 30, carrier_prob = 0.002,
                       RVfounder = TRUE,
                       stop_year = 2015,
                       recall_probs = c(1),
                       founder_byears = c(1900, 1905),
                       FamID = 1)[[2]]
summary(RVped2015)
Ainfo <- affInfo.ped(RVped2015)

#information for the disease-affected relatives
Ainfo[[1]]

# matrix of pairwise kinship coefficents for
# the diseaes-affected relatives
Ainfo[[2]]

data(EgPeds)
egpeds <- new.ped(EgPeds)
summary(egpeds)
Ainfo <- affInfo.ped(ped_file = egpeds)

# Information for the disease-affected
# relatives in the 5 families in egpeds.
Ainfo[[1]]

# matrix of pairwise kinship coefficents for
# the disease-affected relatives in family 4.
Ainfo[[2]][[4]]

}
\references{
Terry M Therneau and Jason Sinnwell (2015). \strong{kinship2: Pedigree Functions.} \emph{R package version 1.6.4.} https://CRAN.R-project.org/package=kinship2
}
\seealso{
\code{\link{new.ped}}
}

