% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeSpecific_Hazards.R
\docType{data}
\name{AgeSpecific_Hazards}
\alias{AgeSpecific_Hazards}
\title{Example age-specific hazards}
\format{A data frame with 100 rows and 3 variables:
\describe{
\item{pop_onset_hazard}{The age-specific population onset hazard}
\item{pop_death_hazard}{The age-specific population death hazard}
\item{affected_death_hazard}{The age-specific population death hazard for affected individuals}
}}
\usage{
data(AgeSpecific_Hazards)
}
\description{
A dataset containing 3 columns of age-specific hazards.  Each age-specific hazard column lists the age-specific hazards in yearly increments, beginning at age 0 ending with age 100.  That is, values in the first row are the age-specific hazards for individuals whose ages fall in the interval [0, 1), the values in the first row are the age-specific hazards for individuals whose ages fall in the interval [1, 2), etc.
}
\keyword{datasets}

