% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManipulatePed_Functions.R
\name{censor_ped}
\alias{censor_ped}
\title{Censor pedigree information after a specified year.}
\usage{
censor_ped(ped_file, censor_year)
}
\arguments{
\item{ped_file}{data.frame. A ped file returned by the sim_RVpedigree function.}

\item{censor_year}{Numeric. The censor year.}
}
\value{
censored_ped The censored pedigree.
}
\description{
The \code{censor_ped} function censors a pedigree relative to a specified year.
}
\details{
censors a pedigree relative to a specified year.  Upon supplying a pedigree and a censor year the The \code{censor_ped} function will remove all individuals born after the censor year, and censor all disease onset events and death events that occured after the censor year
}
\examples{
#Read in example pedigree to trim
data(AgeSpecific_Hazards)


#Specify onset hazard and death hazard
my_onset_hazard <- AgeSpecific_Hazards[,1]
my_death_hazard <- AgeSpecific_Hazards[,c(2,3)]

#Specify age partition.
age_part <- seq(0, 100, by = 1)


set.seed(3)
ex_RVped <- sim_RVpedigree(onset_hazard = my_onset_hazard,
                           death_hazard = my_death_hazard,
                           part = age_part,
                           num_affected = 2,
                           ascertain_span = c(1900, 2015),
                           RR = 10, stop_year = 2015,
                           recall_probs = c(1),
                           founder_byears = c(1900, 1980),
                           FamID = 1)


}

