% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManipulatePed_Functions.R
\name{censor_ped}
\alias{censor_ped}
\title{Censor pedigree information after a specified year.}
\usage{
censor_ped(ped_file, censor_year)
}
\arguments{
\item{ped_file}{data.frame. The pedigree to censor, see details.}

\item{censor_year}{Numeric. The censor year. If missing, when pedigree contains a proband, \code{censor_year} will assume the value of the proband's onset year. See details.}
}
\value{
censored_ped The censored pedigree.
}
\description{
\code{censor_ped} censors a pedigree of any information that occurs after a specified year.
}
\details{
Upon supplying a pedigree and a censor year the \code{censor_ped} function will remove all individuals born after \code{censor_year} and censor all disease onset and death events after the \code{censor_year}.

Users who wish to censor pedigrees that have not been simulated using the \code{SimRVpedigree} package must ensure that the pedigree, \code{ped_file}, supplied to \code{censor_ped} contains the following variables for each pedigree member:

\enumerate{
\item \code{ID}: an identification number.
\item \code{dad_id}: identification number of father.
\item \code{mom_id}: identification number of mother.
\item \code{gender}: gender identification; if male \code{gender = 0}, if female \code{gender = 1}.
\item \code{affected}: affection status, if affected by disease \code{affected  = 1}, otherwise, \code{affected = 0}.
\item \code{birth_year}: the individual's birth year.
\item \code{onset_year}: the individual's disease onset year, when applicable.
\item \code{death_year}: the individual's death year, when applicable.
\item \code{proband}: (Optional) Proband identification variable, \code{proband = 1} if the individual is the proband, and 0 otherwise.
}

If an individual has not experienced disease onset and/or death, then \code{onset_year = NA} and/or \code{death_year = NA}.

If \code{censor_year} is missing, when the pedigree contains a proband, \code{censor_year} is set, internally, to the year that the proband experienced disease onset. However, if \code{ped_file} does not contain the proband identification variable the user must supply a value for \code{censor_year}.
}
\examples{
#Read in example pedigree to trim
data(AgeSpecific_Hazards)


#Specify onset hazard and death hazard
my_onset_hazard <- AgeSpecific_Hazards[,1]
my_death_hazard <- AgeSpecific_Hazards[,c(2,3)]

#Specify age partition.
age_part <- seq(0, 100, by = 1)


#Simulate pedigree
set.seed(3)
ex_RVped <- sim_RVpedigree(onset_hazard = my_onset_hazard,
                           death_hazard = my_death_hazard,
                           part = age_part,
                           num_affected = 2,
                           ascertain_span = c(1900, 2015),
                           RR = 10, stop_year = 2015,
                           recall_probs = c(1),
                           founder_byears = c(1900, 1980),
                           FamID = 1)

library(kinship2)
Original_ped <- with(ex_RVped[[2]], pedigree(id = ID,
                                             dadid = dad_id,
                                             momid = mom_id,
                                             sex = gender + 1,
                                             affected = affected))
plot(Original_ped)

Censor_ped1 <- with(censor_ped(ex_RVped[[2]]),
                    pedigree(id = ID,
                             dadid = dad_id,
                             momid = mom_id,
                             sex = gender + 1,
                             affected = affected))
plot(Censor_ped1)

Censor_ped2 <- with(censor_ped(ex_RVped[[2]], 2000),
                    pedigree(id = ID,
                             dadid = dad_id,
                             momid = mom_id,
                             sex = gender + 1,
                             affected = affected))
plot(Censor_ped2)

}

