% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManipulatePed_Functions.R
\name{censor_ped}
\alias{censor_ped}
\title{Censor Pedigree After a Specified Year}
\usage{
censor_ped(ped_file, censor_year)
}
\arguments{
\item{ped_file}{Data.frame. The pedigree to censor, see details.}

\item{censor_year}{Numeric. The censor year. If missing, when pedigree contains a proband, \code{censor_year} will assume the value of the proband's onset year. See details.}
}
\value{
censored_ped The censored pedigree.
}
\description{
\code{censor_ped} censors a pedigree of any information that occurs after a specified year.
}
\details{
Upon supplying a pedigree and a censor year the \code{censor_ped} function will remove all individuals born after \code{censor_year} and censor all disease onset and death events after the \code{censor_year}.

Users who wish to censor pedigrees which have not been simulated by \code{\link{sim_RVped}} or \code{\link{sim_ped}} must ensure that the pedigree, \code{ped_file}, supplied to \code{censor_ped} contains the following variables for each pedigree member:

\enumerate{
\item \code{ID}: an identification number.
\item \code{dadID}: identification number of father.
\item \code{momID}: identification number of mother.
\item \code{sex}: sex identification; if male \code{sex = 0}, if female \code{sex = 1}.
\item \code{affected}: affection status, if affected by disease \code{affected  = 1}, otherwise, \code{affected = 0}.
\item \code{birthYr}: the individual's birth year.
\item \code{onsetYr}: the individual's disease onset year, when applicable.
\item \code{deathYr}: the individual's death year, when applicable.
\item \code{proband}: (Optional) Proband identification variable, \code{proband = 1} if the individual is the proband, and 0 otherwise.
}

If an individual has not experienced disease onset and/or death, then \code{onsetYr = NA} and/or \code{deathYr = NA}.

If \code{censor_year} is missing, when the pedigree contains a proband, \code{censor_year} is set, internally, to the year that the proband experienced disease onset. However, if \code{ped_file} does not contain the proband identification variable the user must supply a value for \code{censor_year}.

For a detailed example please refer to vignette.
}
\examples{
#Read in example pedigree to trim
data(AgeSpecific_Hazards)

#Simulate a pedigree ascertained for multiple affecteds
set.seed(3)
ex_RVped <- sim_RVped(onset_hazard = AgeSpecific_Hazards[,1],
                      death_hazard = AgeSpecific_Hazards[,c(2,3)],
                      part = seq(0, 100, by = 1),
                      num_affected = 2,
                      ascertain_span = c(1900, 2015),
                      RR = 10, stop_year = 2015,
                      recall_probs = c(1),
                      founder_byears = c(1900, 1980),
                      FamID = 1)[[2]]

#To plot pedigrees, use the kinship2 package
library(kinship2)

#assign simulated pedigree to pedigree object, then pass to plot function
Original_ped <- pedigree(id = ex_RVped$ID,
                         dadid = ex_RVped$dadID,
                         momid = ex_RVped$momID,
                         sex = ex_RVped$sex + 1,
                         affected = ex_RVped$affected)
plot(Original_ped)

#Now censor ex_RVped after 1999
Cped <- censor_ped(ped_file = ex_RVped, censor_year = 1999)

#Assign the the censored pedigree to a pedigree object and plot.
Censor_ped <- pedigree(id = Cped$ID,
                       dadid = Cped$dadID,
                       momid = Cped$momID,
                       sex = Cped$sex + 1,
                       affected = Cped$affected)
plot(Censor_ped)


}
