% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_and_warning_functions.R
\name{check_dhaz}
\alias{check_dhaz}
\title{Check to see if death_hazard is appropriately specified}
\usage{
check_dhaz(death_hazard)
}
\arguments{
\item{death_hazard}{data.frame. The age-specific death hazards for affected and unaffected individuals.}
}
\description{
Check to see if death_hazard is appropriately specified
}

