% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_and_warning_functions.R
\name{check_hazpart}
\alias{check_hazpart}
\title{Check to see if hazard and partition vectors are specified correctly}
\usage{
check_hazpart(hazard, part)
}
\arguments{
\item{hazard}{A numeric vector.  A vector of age-specific hazards.}

\item{part}{A numeric vector. The partition over which to apply hazard, note length(part) == length(hazard) + 1 should return TRUE.}
}
\description{
\code{check_hazpart} Check to see if hazard and partition vectors are specified correctly
}
\examples{
check_hazpart(hazard = c(0.1, 0.4, 0.7, 0.8, 0.9), part = c(0, 20, 40, 60, 80, 100))
check_hazpart(hazard = c(0.1, 0.4, 0.7, 0.8, 0.9), part = c(0, 20, 40, 60, 80))
}

