% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_and_warning_functions.R
\name{check_part}
\alias{check_part}
\title{Check that age-specific hazards span an appropriate range of ages.}
\usage{
check_part(part)
}
\arguments{
\item{part}{A numeric vector.  Partition of ages over which to apply the
age-specific hazards.}
}
\description{
Check that age-specific hazards span an appropriate range of ages.
}
\examples{
\dontrun{check_part(part = seq(10, 20, by = 1))}
\dontrun{check_part(part = seq(-10, 10, by = 1))}
check_part(part = seq(0, 85, by = 1))
}

