% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedCheck_functions.R
\name{check_ped}
\alias{check_ped}
\title{Checks ped_file input for expected information.}
\usage{
check_ped(ped_file)
}
\arguments{
\item{ped_file}{Data.frame. A pedigree generated by \code{sim_ped}.}
}
\description{
Checks ped_file input for expected information.
}
\keyword{internal}
