% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_DF_Ped.R
\name{check_ped}
\alias{check_ped}
\title{Checks ped_file for expected information, used before converting to ped object.}
\usage{
check_ped(ped_file)
}
\arguments{
\item{ped_file}{data.frame The pedigree in data frame format.}
}
\description{
Checks ped_file for expected information, used before converting to ped object.
}
\keyword{internal}
