% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedSim_Functions.R
\name{create_mate}
\alias{create_mate}
\title{Create a new mate}
\usage{
create_mate(partner_info, last_id, GRR, carrier_prob, RVfounder, intro_RV)
}
\arguments{
\item{partner_info}{Data.frame with 1 row. All pedigree information for the individual requiring a new mate, must contain all fields generated by \code{\link{create_pedFile}}.}

\item{last_id}{Numeric.  The last id used in the pedigree.}

\item{GRR}{Numeric. The relative-risk of disease for individuals who inherit the causal variant.}

\item{carrier_prob}{Numeric.  The population allele frequency of the genetic variant with relative risk \code{GRR}. See details.}

\item{RVfounder}{String. \code{RVfounder} may take on one of three values: \code{"single"}, \code{"first"}, or \code{"multiple"}.  If "first" the first founder will introduce the rare variant to the pedigree, if "single" at most one founder will be able to intoduce the rare variant to the pedigree, if "multiple" more than one founder can intrduce the variant to the pedigree.}

\item{intro_RV}{Logical. If \code{intro_RV = TRUE}, then at least 1 founder has introduced the causal variant to the pedigree. If \code{intro_RV = FALSE} then the variant has not yet been introduced to the pedigree.}
}
\value{
\code{new_mate_info} the pedigree information for the new mate.

\code{last_id} the updated value of last_id after adding the new mate.

\code{RV_founder} a true/false value indicating if at least 1 founder has introduced the RV.
}
\description{
Create new mating partner in pedigree.
}
\keyword{internal}
