% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedSim_Functions.R
\name{create_mate}
\alias{create_mate}
\title{Create a new mate}
\usage{
create_mate(partner_info, last_id, GRR, allele_freq, RVfounder, intro_RV)
}
\arguments{
\item{partner_info}{Data.frame with 1 row. All pedigree information for the individual requiring a new mate, must contain all fields generated by \code{\link{create_pedFile}}.}

\item{last_id}{Numeric.  The last id used in the pedigree.}

\item{GRR}{Numeric. The genetic relative-risk of disease, i.e. the relative-risk of disease for individuals who carry at least one copy of the causal variant.}

\item{allele_freq}{Numeric.  The population allele frequency of variants with relative risk \code{GRR}. See details.}

\item{RVfounder}{String. \code{RVfounder} may take on one of three values: \code{"single"}, \code{"first"}, or \code{"multiple"}.  See details.}

\item{intro_RV}{Logical. If \code{intro_RV = TRUE}, then at least 1 founder has introduced the causal variant to the pedigree. If \code{intro_RV = FALSE} then the variant has not yet been introduced to the pedigree.}
}
\value{
\code{new_mate_info} the pedigree information for the new mate.

\code{last_id} the updated value of last_id after adding the new mate.

\code{RV_founder} a true/false value indicating if at least 1 founder has introduced the RV.
}
\description{
Create new mating partner in pedigree.
}
\keyword{internal}
