% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedSim_Functions.R
\name{create_mate}
\alias{create_mate}
\title{Create a new mate}
\usage{
create_mate(partner_info, last_id, carrier_prob, RVfounder)
}
\arguments{
\item{partner_info}{Data.frame with 1 row. All pedigree information for the individual requiring a new mate, must contain all fields generated by \code{\link{create_pedFile}}.}

\item{last_id}{Numeric.  The last id used in the pedigree.}

\item{carrier_prob}{Numeric.  The carrier probability for all causal variants with relative-risk of disease \code{GRR}.  By default, \code{carrier_prob}\code{ = 0.002}}

\item{RVfounder}{Logical.  Indicates if all pedigrees segregate the rare, causal variant.  By default, \code{RVfounder = FALSE} See details.}
}
\value{
\code{new_mate_info} the pedigree information for the new mate.

\code{last_id} the updated value of last_id after adding the new mate.
}
\description{
Create new mating partner in pedigree.
}
\keyword{internal}
