% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedSim_Functions.R
\name{create_pedFile}
\alias{create_pedFile}
\title{Create empty ped file}
\usage{
create_pedFile()
}
\value{
A empty data frame with all the necessary fields needed to generate
a ped file using SimRVPedigree
}
\description{
\code{create_pedFile} initializes an empty data frame with all of the appropriate fields to simulate a pedigree using SimRVPedigree.
}

