% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedSim_Functions.R
\name{create_pedFile}
\alias{create_pedFile}
\title{Create an empty pedigree}
\usage{
create_pedFile()
}
\value{
An empty data frame with all fields required by \code{simPed} to generate a pedigree.
}
\description{
\code{create_pedFile} initializes an empty data frame with all of the fields required by \code{simPed} to simulate a pedigree.
}
\keyword{internal}
