% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHPP_functions.R
\name{findWaitProb}
\alias{findWaitProb}
\title{computes the probability that a waiting time, W, is less than or equal to some value s, given that the last event occured at s_0}
\usage{
findWaitProb(last_event, wait_time, hazard, part, scale = FALSE)
}
\arguments{
\item{last_event}{A numeric constant.  The age at last event.}

\item{wait_time}{A numeric constant. The wiating time, in years, to next event.}

\item{hazard}{A numeric vector.  A vector of age-specific hazards.}

\item{part}{A numeric vector.  Partition of ages over which to apply the age-specific hazards.}

\item{scale}{Logical. By default scale = FALSE.  NOTE: Choosing scale = TRUE ensures that W is a proper random variable, i.e. that this function is a proper CDF with upper limit 1.}
}
\value{
wait_prob numeric. The probability that the waiting time is at most wait_time given that the last event occured at last_event
}
\description{
computes the probability that a waiting time, W, is less than or equal to some value s, given that the last event occured at s_0
}
\examples{
findWaitProb(last_event = 2, wait_time = 18, hazard = seq(1, 2.5, by = 0.1), part = seq(0, 80, by = 5))
findWaitProb(last_event = 2, wait_time = 18, hazard = seq(0, 1.5, by = 0.1), part = seq(0, 80, by = 5), scale = TRUE)
findWaitProb(last_event = 2, wait_time = 18, hazard = seq(0, 1.5, by = 0.1), part = seq(0, 80, by = 5))

}

