% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManipulatePed_Functions.R
\name{get_famInfo}
\alias{get_famInfo}
\title{Obtain summary information}
\usage{
get_famInfo(ped_file)
}
\arguments{
\item{ped_file}{}
}
\value{
A data frame containing the relevant summary information
}
\description{
Obtain summary information
}
\keyword{internal}
