% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeSim_Functions.R
\name{get_lifeEvents}
\alias{get_lifeEvents}
\title{Simulate all life events.}
\usage{
get_lifeEvents(onset_hazard, death_hazard, part, birth_range, NB_params, RR,
  YOB)
}
\arguments{
\item{onset_hazard}{Numeric vector. The population age-specific onset hazard.}

\item{death_hazard}{data.frame. Column 1 should specify the age specific
mortality rates in the unaffected population, while column 2 should provide
the age specific morality rates in the affected population.}

\item{part}{A numeric vector.  Partition of ages over which to apply the
age-specific hazards}

\item{birth_range}{A numeric vector of length 2.  The minimum and maximum
allowable birth ages in simulation.}

\item{NB_params}{A numeric vector of length 2. The size and probabiliy
parameters of the negative binomial distribution that describes the number of
children per household in the population.}

\item{RR}{A numeric constant. The relative risk of developing
disease for individuals who have inherited the rare variant.}

\item{YOB}{numeric. The indivdiual's year of birth.}
}
\value{
Named matrix. The waiting times between all life events simulated for an individual, named according to which life event has occurred.
}
\description{
\code{get_lifeEvents} simulates all life events for an individual starting at age
0 and ending with death by applying the \link{get_nextEvent} function until death
occurs.
}
\examples{
part_vec <- seq(0, 100, by = 1)
unaffected_mort <- 0.00001 + pgamma(seq(0.16, 16, by = .16),
                                    shape = 9.5, scale = 1)
affected_mort <- c(0.55, 0.48, 0.37, 0.23, 0.15,
                   pgamma(seq(0.96, 16, by = .16), shape = 4, scale = 1.5))
Dhaz_df  <- (as.data.frame(cbind(unaffected_mort, affected_mort)))
Ohaz_vec <- (dgamma(seq(0.1, 10, by = .1), shape = 8, scale = 0.75))
set.seed(17)
get_lifeEvents(onset_hazard = Ohaz_vec,
               death_hazard = Dhaz_df,
               part = part_vec,
               birth_range = c(17,45),
               NB_params = c(2, 4/7), RR = 1,
               YOB = 1900)

set.seed(17)
get_lifeEvents(onset_hazard = Ohaz_vec,
               death_hazard = Dhaz_df,
               part = part_vec,
               birth_range = c(17,45),
               NB_params = c(2, 4/7), RR = 15,
               YOB = 1900)

}

