% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeSim_Functions.R
\name{get_lifeEvents}
\alias{get_lifeEvents}
\title{Simulate all life events.}
\usage{
get_lifeEvents(onset_hazard, death_hazard, part, birth_range, NB_params, RR,
  YOB)
}
\arguments{
\item{onset_hazard}{Numeric. The population age-specific hazard rate for disease onset.}

\item{death_hazard}{Data.frame. Column 1 should specify the unaffected age-specific hazard rate for death, and column 2 should specify the affected age-specific hazard rate for death.}

\item{part}{Numeric. The partition of ages over which to apply the age-specific hazard rates in \code{onset_hazard} and \code{death_hazard}.}

\item{birth_range}{Numeric list of length 2. By default, \code{birth_range = c(18, 45)}. The minimum and maximum allowable ages that individuals will be able to reproduce.}

\item{NB_params}{Numeric list of length 2. By default, \code{NB_params = c(2, 4/7)}. The size and probability parameters of the negative binomial distribution used to model the number of children per household.}

\item{RR}{Numeric. The relative risk of disease onset for individuals who have inherited the rare variant.}

\item{YOB}{A positive number. The indivdiual's year of birth.}
}
\value{
A named matrix. A matrix containing the years of all life events simulated for an individual, named by event type.
}
\description{
\code{get_lifeEvents} simulates all life events for an individual starting at age
0 and ending with death through recursive application of \link{get_nextEvent}.
}
\examples{
data(AgeSpecific_Hazards)

my_part <- seq(0, 100, by = 1)
my_onset_haz <- AgeSpecific_Hazards[,1]
my_death_haz <- AgeSpecific_Hazards[,c(2,3)]

set.seed(17)
get_lifeEvents(onset_hazard = my_onset_haz,
               death_hazard = my_death_haz,
               part = my_part,
               birth_range = c(17,45),
               NB_params = c(2, 4/7), RR = 1,
               YOB = 1900)

set.seed(17)
get_lifeEvents(onset_hazard = my_onset_haz,
               death_hazard = my_death_haz,
               part = my_part,
               birth_range = c(17,45),
               NB_params = c(2, 4/7), RR = 15,
               YOB = 1900)

}

