% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeSim_Functions.R
\name{get_nextEvent}
\alias{get_nextEvent}
\title{Simulate next life event.}
\usage{
get_nextEvent(current_age, disease_status, lambda_birth, onset_hazard,
  death_hazard, part, birth_range, RR)
}
\arguments{
\item{current_age}{numeric. The individual's current age.}

\item{disease_status}{numeric. The individual's disease status, set disease_status = 1 if individual has experienced disease onset, otherwise set disease_status = 0.}

\item{lambda_birth}{numeric. The individual's birth rate.}

\item{onset_hazard}{Numeric. The population age-specific hazard rate for disease onset.}

\item{death_hazard}{Data.frame. Column 1 should specify the unaffected age-specific hazard rate for death, and column 2 should specify the affected age-specific hazard rate for death.}

\item{part}{Numeric. The partition of ages over which to apply the age-specific hazard rates in \code{onset_hazard} and \code{death_hazard}.}

\item{birth_range}{Numeric list of length 2. By default, \code{birth_range = c(18, 45)}. The minimum and maximum allowable ages that individuals will be able to reproduce.}

\item{RR}{Numeric. The relative risk of disease onset for individuals who have inherited the rare variant.}
}
\value{
A named matrix. The number of years until the next life event,
named by event type.  See Details.
}
\description{
\code{get_nextEvent} randomly simulates the next life event for an individual
 given their current age, disease status, and relative risk of disease onset.
}
\details{
\code{get_nextEvent} randomly simulates the next life event for an individual by
generating the waiting times, via \link{get_WaitTime}, to reproduction, onset,
and death given the individual's current age.  The event with the shortest
waiting time is chosen as the next life event.

If get_nextEvent returns a value named:
 - "Birth" the next life event is reproduction,
 - "Onset" the next life event is disease onset,
 - "Death" the next life event is death
}
\examples{
data(AgeSpecific_Hazards)

my_part <- seq(0, 100, by = 1)
my_onset_haz <- AgeSpecific_Hazards[,1]
my_death_haz <- AgeSpecific_Hazards[,c(2,3)]

set.seed(17)
get_nextEvent(current_age = 23, disease_status = 0,
              lambda_birth = 0.05,
              onset_hazard = my_onset_haz,
              death_hazard = my_death_haz,
              part = my_part,
              birth_range = c(17,45), RR = 5)

set.seed(17)
get_nextEvent(current_age = 23, disease_status = 1,
              lambda_birth = 0.05,
              onset_hazard = my_onset_haz,
              death_hazard = my_death_haz,
              part = my_part,
              birth_range = c(17,45), RR = 5)

}

