% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHPP_functions.R
\name{get_WaitTime}
\alias{get_WaitTime}
\title{Simulate the waiting time to next event for a non-homogeneous Poisson process}
\usage{
get_WaitTime(p, last_event, hazard, part, scale = FALSE)
}
\arguments{
\item{p}{A numeric constant. Argument of non-homogeneous poisson process quantile function}

\item{last_event}{A numeric constant.  The age at last event.}

\item{hazard}{A numeric vector.  A vector of age-specific hazards.}

\item{part}{A numeric vector.  Partition of ages over which to apply the
age-specific hazards.}

\item{scale}{Logical. By default scale = FALSE.  Specifying scale = TRUE ensures that W is a proper random variable, i.e. that this function is a proper CDF with upper limit 1.}
}
\value{
The waiting time to next event, units same as those in \code{part}.
}
\description{
\code{get_WaitTime} simulates the waiting time to next event for a non-homogeneous Poisson process
}
\details{
\code{get_WaitTime} simulates the waiting time to next event for a non-homogeneous Poisson process.  The units of the simulated waiting time are the units specified in \code{part}, i.e. if \code{part} is specified in years, the simulated waiting time is in years.  Makes use of \link{approxfun} and \link{splinefun}
}
\examples{
haz_vec <- c(seq(0, 0.5, by = 0.05), rev(seq(0.46, 0.5, by = 0.01)))
part_vec <- seq(0, 80, by = 5)

get_WaitTime(p = 0.05, last_event = 0,
             hazard = haz_vec, part = part_vec,
             scale = FALSE)

get_WaitTime(p = 0.05, last_event = 0,
             hazard = haz_vec, part = part_vec,
             scale = TRUE)

}

