% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hazard_object_functions.R
\name{hazard}
\alias{hazard}
\title{Constructor function for an object of class \code{hazard}}
\usage{
hazard(hazardDF, partition)
}
\arguments{
\item{hazardDF}{Data.frame. Column 1 specifies the age-specific hazard rate of \emph{disease} in the population of interest, column 2 specifies the age-specific hazard rate for \emph{death} in the \strong{unaffected} population, and column 3 specifies the age-specific hazard rate for \emph{death} in the \strong{affected} population.  See details.}

\item{partition}{Numeric vector. Optional. The partition of ages, in years, over which to apply the age-specific hazard rates in \code{hazardDF}. If missing, it is assumed that \code{partition} starts at 0 and increases in yearly increments.  See details.}
}
\value{
an object of class \code{hazard}.
}
\description{
Constructor function for an object of class \code{hazard}
}
\examples{
data(AgeSpecific_Hazards)
my_HR <- hazard(partition = seq(0, 100, by = 1),
                hazardDF = AgeSpecific_Hazards)
class(my_HR)

}
