% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hazard_object_functions.R
\name{hazard}
\alias{hazard}
\title{Constructor function of object of class \code{hazard}}
\usage{
hazard(partition, hazardDF)
}
\arguments{
\item{partition}{Numeric vector. The partition of ages over which to apply the age-specific hazard rates in \code{hazardDF}.}

\item{hazardDF}{Data.frame. Column 1 should specify the age-specific hazard rate for individuals with no genetic predisposition to the disease, column 2 should specify the age-specific hazard rate for death in the unaffected population, and column 3 should specify the age-specific hazard rate for death in the affected population.}
}
\value{
an object of class \code{hazard}.
}
\description{
Constructor function of object of class \code{hazard}
}
\examples{
my_HR <- hazard(partition = seq(0, 100, by = 1),
                hazardDF = AgeSpecific_Hazards)
class(my_HR)

}
