% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ped_object_functions.R
\name{is.ped}
\alias{is.ped}
\title{Check to see if object is of class ped}
\usage{
is.ped(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
Logical. Indicates if \code{x} is of class \code{ped}.
}
\description{
Check to see if object is of class ped
}
\keyword{internal}
