% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ped_Methods.R
\name{ped}
\alias{ped}
\title{Constructor function for an object of class ped}
\usage{
ped(ped_file)
}
\arguments{
\item{ped_file}{A pedigree generated by \code{sim_ped} pr \code{sim_RVped}.}
}
\value{
an object of class \code{ped}.
}
\description{
Constructor function for an object of class ped
}
\keyword{internal}
