% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ped2Pedigree_functions.R
\name{ped2pedigree}
\alias{ped2pedigree}
\title{Convert a ped object to a kinship2 pedigree object}
\usage{
ped2pedigree(x)
}
\arguments{
\item{x}{A ped object.}
}
\value{
A pedigree object.  See ?kinship2::pedigree for details.
}
\description{
Convert a SimRVPedigree ped object to a kinship 2 pedigree object
}
\examples{
data(EgPeds)

ped1 = new.ped(EgPeds[EgPeds$FamID == 1, ])
head(ped1, n = 3)
class(ped1)
summary(ped1)

kinPed <- ped2pedigree(ped1)
kinPed

library(kinship2)
plot(kinPed)
pedigree.legend(kinPed, location="topleft", radius=.25)



AllPeds = new.ped(EgPeds)
summary(AllPeds)

kinPed_multi <- ped2pedigree(AllPeds)
kinPed_multi

unique(AllPeds$FamID)

# If pedigree object contains more than 1 family
# take care to specify the family ID when plotting
plot(kinPed_multi['2'])
pedigree.legend(kinPed_multi['2'], location="topleft", radius=.25)

plot(kinPed_multi['4'])
pedigree.legend(kinPed_multi['4'], location="topleft", radius=.25)



}
