% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ped2Pedigree_functions.R
\name{ped2pedigree}
\alias{ped2pedigree}
\title{Create a kinship2 pedigree structure from an object of class ped}
\usage{
ped2pedigree(x)
}
\arguments{
\item{x}{A ped object.}
}
\value{
A pedigree object.  See \code{\link{pedigree}} for details.
}
\description{
Create a kinship2 pedigree structure from an object of class ped
}
\references{
Terry M Therneau and Jason Sinnwell (2015). \strong{kinship2: Pedigree Functions.} \emph{R package version 1.6.4.} https://CRAN.R-project.org/package=kinship2
}
