% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ped2Pedigree_functions.R
\name{pedigreeLabels}
\alias{pedigreeLabels}
\title{Create pedigree labels}
\usage{
pedigreeLabels(x, ref_year)
}
\arguments{
\item{x}{An object of class ped.}

\item{ref_year}{A numeric constant.  The reference year used to determine current age for pedigree members.}
}
\value{
A list of labels that can be used with kinship2's plot function. See \code{\link{plot.pedigree}} for details.
}
\description{
Create pedigree labels
}
\examples{
data(EgPeds)

ped1 = new.ped(EgPeds[EgPeds$FamID == 2, ])
head(ped1, n = 3)
class(ped1)
summary(ped1)

kinPed <- ped2pedigree(ped1)
kinPed
kinLabs <- pedigreeLabels(ped1, ref_year = 2016)

library(kinship2)
plot(kinPed,
     id = kinLabs,
     mar = c(3, 3, 4, 3) + 0.1)
pedigree.legend(kinPed, location = "topleft", radius = 0.2)


AllPeds = new.ped(EgPeds)
summary(AllPeds)

kinPed_multi <- ped2pedigree(AllPeds)
kinPed_multi
kinLabs_multi <- pedigreeLabels(AllPeds, ref_year = 2016)

unique(AllPeds$FamID)

# If pedigree object contains more than 1 family
# take care to specify the family ID when plotting
plot(kinPed_multi['4'],
     id = kinLabs_multi[AllPeds$FamID == 4],
     mar = c(3, 3, 4, 3) + 0.1)
pedigree.legend(kinPed['4'], location = "topleft", radius = 0.2)


}
\references{
Terry M Therneau and Jason Sinnwell (2015). \strong{kinship2: Pedigree Functions.} \emph{R package version 1.6.4.} https://CRAN.R-project.org/package=kinship2
}

