% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ped_object_functions.R
\name{plot.ped}
\alias{plot.ped}
\title{Plot pedigree}
\usage{
\method{plot}{ped}(x, ref_year, legendLocation = "topleft",
  legendRadius = 0.25, ...)
}
\arguments{
\item{x}{An object of class ped.}

\item{ref_year}{When provided, the reference year for age age labels.  Users may supply a (numeric) year which will create age labels at the specified year. Alternatively, users may set \code{ref_year}\code{ = "ascYR"}, which will create age lables for the year the pedigree was ascertained, when ascertained.  When missing, no age labels are created.}

\item{legendLocation}{The location for the pedigree legend, \code{"topleft"}, \code{"topright"}, \code{"bottomright"}, \code{"bottomleft"}.  By default, \code{legendLocation}\code{ = "topleft"}.}

\item{legendRadius}{The radius of the pedigree legend.  By default, \code{legendRadius}\code{ = 0.25}.}

\item{...}{Extra options that feed to \code{\link{plot.pedigree}}, or \code{\link{plot}}.}
}
\description{
Plot pedigree
}
\examples{
#Read in age-specific harard data and create hazard object.
data(AgeSpecific_Hazards)
haz_obj <- hazard(hazardDF = AgeSpecific_Hazards)

#Simulate a pedigree ascertained for multiple affecteds
set.seed(4)
RVped2015 <- simRVped(hazard_rates = haz_obj,
                       num_affected = 2,
                       ascertain_span = c(1900, 2015),
                       GRR = 30, carrier_prob = 0.002,
                       RVfounder = TRUE,
                       stop_year = 2015,
                       recall_probs = c(1),
                       founder_byears = c(1900, 1925),
                       FamID = 1)[[2]]
summary(RVped2015)

#plot pedigree without age labels
plot(RVped2015, cex = 0.25)

#plot pedigree with age labels, since unspecified
#reference year defaults to ascertainment year
plot(RVped2015, ref_year = "ascYR")
plot(RVped2015, ref_year = "ascYR", cex= 0.75, symbolsize = 1.25)
plot(RVped2015, ref_year = 2000,
     cex= 1, symbolsize = 1.5,
     legendLocation = "topleft",
     legendRadius = 0.1)

#plot pedigree with age lablels at specified reference years.
plot(RVped2015, ref_year = 2015,
cex= 0.75, symbolsize = 1.25, legendRadius = 0.15)
plot(RVped2015, ref_year = 2005, cex= 0.75, symbolsize = 1.25)
plot(RVped2015, ref_year = 1995, cex= 0.75, symbolsize = 1.25)
plot(RVped2015, ref_year = 1985, cex= 0.75, symbolsize = 1.25)
plot(RVped2015, ref_year = 1975, cex= 0.75, symbolsize = 1.25)
plot(RVped2015, ref_year = 1960, cex= 0.75, symbolsize = 1.25)

}
\seealso{
\code{\link{plot.pedigree}}, \code{\link{plot}}
}

