% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeSim_Functions.R
\name{sim_lifeEvents}
\alias{sim_lifeEvents}
\title{Simulate all life events}
\usage{
sim_lifeEvents(hazard_rates, birth_range, NB_params, RR, YOB)
}
\arguments{
\item{hazard_rates}{on object of class \code{hazard}, created by \code{\link{new.hazard}}.}

\item{birth_range}{Numeric vector of length 2. The minimum and maximum allowable ages, in years, between which individuals may reproduce.  By default, \code{birth_range = c(18, 45)}.}

\item{NB_params}{Numeric vector of length 2. The size and probability parameters of the negative binomial distribution used to model the number of children per household.  By default, \code{NB_params = c(2, 4/7)}, due to the investigation of Kojima and Kelleher (1962).}

\item{RR}{Numeric. The individual's relative-risk of disease.}

\item{YOB}{A positive number. The indivdiual's year of birth.}
}
\value{
A named matrix containing the years of an individual's simulated life events, named by event type, see details.
}
\description{
Primarily intended as an internal function, \code{sim_lifeEvents} simulates all life events for an individual starting at birth, age 0, and ending with death.
}
\details{
Starting at birth, age 0, the \code{sim_lifeEvents} generates waiting times to reproduction, onset, and death. The event with the shortest waiting time is chosen as the next life event, and individual's age is updated by the waiting time of the winning event.  Conditioned on the individual's new age, this process is applied recursively, until death occurs and the process terminates.

We make the following assumptions regarding the simulation of waiting times:
 \enumerate{
 \item We assume that, given an individual's current age, their time to disease onset is the waiting time in a non-homogeneous Poisson process with an age-specific hazard rate that follows a proportional hazards model.  In this model, individuals who have NOT inherited the rare variant experience disease onset according to the baseline (or population) hazard rate of disease.  On the other hand, individuals who have inherited the rare variant are assumed to have an increased risk of disease onset relative to those who have inherited it.  The user is expected to supply the baseline hazard rate of disease, as well as the relative-risk of disease for genetic cases. Additionally, we impose the restriction that individuals may only experience disease onset once, and remain affected from that point on.

\item We assume that, given an individual's current age, their time to death is the waiting time in a non-homogeneous Poisson process with age-specific hazard rate determined by their affection status.  We assume that unaffected individuals experience death according to the age-specific hazard rate for death in the unaffected population; if the disease of interest is sufficiently rare, the user may instead choose to substitute the population age-specific hazard rate for death in the general population.  We assume that affected individuals experience death according to the age-specific hazard rate for death in the affected population.  The user is expected to supply both of these age-specific hazard rates.

\item We assume that, given an individual's current age, their time to reproduction is the waiting time in a homogeneous Poisson process.  That is, we assume that individuals reproduce at uniform rate during their reproductive years.  For example, one's reproductive years may span from age 18 to age 45 years.  We do not allow for offspring to be produced outside of an individual's \code{birth_range}.

}

\code{sim_lifeEvents} will return a named matrix, which contains the years of the simulated life events, named by event type.  The possible event types are as follows:
\itemize{
 \item "Child" a reproductive event, i.e. creation of offspring
 \item "Onset" disease onset event,
 \item "Death" death event
}
}
\examples{
data(AgeSpecific_Hazards)

my_HR <- new.hazard(partition = seq(0, 100, by = 1),
                    hazardDF = AgeSpecific_Hazards)

# The following commands simulate all life events for an individual, whose
# relative-risk of disease is 1, born in 1900.  From the output, this
# individual has 1 child in 1944, and then dies in 1961.
set.seed(1234)
sim_lifeEvents(hazard_rates = my_HR,
               birth_range = c(17,45),
               NB_params = c(2, 4/7), RR = 1,
               YOB = 1900)

# Using the same random seed, notice how the life events can vary for
# someone with an increased relative-risk of disease, say 25.
# From the output, this individual also has 1 child in 1944, then
# experiences disease onset in 1949, and dies in 1957.
set.seed(1234)
sim_lifeEvents(hazard_rates = my_HR,
               birth_range = c(17,45),
               NB_params = c(2, 4/7), RR = 25,
               YOB = 1900)

}
\references{
OUR MANUSCRIPT

Ken-Ichi Kojima, Therese M. Kelleher. (1962), \emph{Survival of Mutant Genes}. The American Naturalist 96, 329-346.
}
