% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedSim_Functions.R
\name{sim_nFam}
\alias{sim_nFam}
\title{Simulate a nuclear family from a founder.}
\usage{
sim_nFam(found_info, stop_year, last_id, onset_hazard, death_hazard, part,
  birth_range, NB_params, RR)
}
\arguments{
\item{found_info}{a data.frame with 1 row. All ped file information for the founder, must contain all fields generated by \code{\link{create_pedFile}}.}

\item{stop_year}{Numeric. The last year of study.  If missing, the current year is used.}

\item{last_id}{numeric.  The last id used in the pedigree.}

\item{onset_hazard}{Numeric. the population age-specific hazard rate for disease.}

\item{death_hazard}{Data.frame. Column 1 should specify the age-specific hazard rate for death in the unaffected population, and column 2 should specify the age-specific hazard rate for death in the affected population. See details.}

\item{part}{Numeric. The partition of ages over which to apply the age-specific hazard rates in \code{onset_hazard} and \code{death_hazard}.}

\item{birth_range}{Numeric list of length 2. The minimum and maximum allowable ages between which individuals may reproduce.  If missing, it is assumed that \code{birth_range = c(18, 45)}.}

\item{NB_params}{Numeric list of length 2. The size and probability parameters of the negative binomial distribution used to model the number of children per household.  If missing, it is assumed that \code{NB_params = c(2, 4/7)}}

\item{RR}{Numeric. The relative risk of disease for individuals who have inherited the rare variant.}
}
\value{
A ped file with all updated life events for the founder, and
additional information for mate and offspring, when offspring are generated.
}
\description{
Simulate all life events for a given individual and create a nuclear family, when
appropriate.
}

