% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedSim_Functions.R
\name{sim_nFam}
\alias{sim_nFam}
\title{Simulate a Nuclear Family from a Single Founder}
\usage{
sim_nFam(found_info, stop_year, last_id, onset_hazard, death_hazard, part,
  birth_range, NB_params, RR)
}
\arguments{
\item{found_info}{Data.frame with 1 row. All pedigree information for the founder, must contain all fields generated by \code{\link{create_pedFile}}.}

\item{stop_year}{Numeric. The last year of study.  If missing, the current year is used.}

\item{last_id}{Numeric.  The last id used in the pedigree.}

\item{onset_hazard}{Numeric. The population age-specific hazard rate for disease.}

\item{death_hazard}{Data.frame. Column 1 should specify the age-specific hazard rate for death in the unaffected population, and column 2 should specify the age-specific hazard rate for death in the affected population. See details.}

\item{part}{Numeric. The partition of ages over which to apply the age-specific hazard rates in \code{onset_hazard} and \code{death_hazard}.}

\item{birth_range}{Numeric list of length 2. The minimum and maximum allowable ages, in years, between which individuals may reproduce.  By default, \code{birth_range = c(18, 45)}.}

\item{NB_params}{Numeric list of length 2. The size and probability parameters of the negative binomial distribution used to model the number of children per household.  By default, \code{NB_params = c(2, 4/7)}, due to the investigation of Kojima and Kelleher (1962).}

\item{RR}{Numeric. The relative risk of disease for individuals who inherit the rare variant.}
}
\value{
A pedigree with updated life events for the founder, and
additional information for mate and offspring, when offspring are generated.
}
\description{
Simulate all life events for an individual and create a nuclear family, when appropriate.
}
\keyword{internal}

