% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedSim_Functions.R
\name{sim_ped}
\alias{sim_ped}
\title{Simulate a pedigree.}
\usage{
sim_ped(onset_hazard, death_hazard, part, RR, FamID, founder_byears, stop_year,
  birth_range = c(18, 45), NB_params = c(2, 4/7))
}
\arguments{
\item{onset_hazard}{Numeric. The population age-specific hazard rate for disease onset.}

\item{death_hazard}{Data.frame. Column 1 should specify the unaffected age-specific hazard rate for death, and column 2 should specify the affected age-specific hazard rate for death.}

\item{part}{Numeric. The partition of ages over which to apply the age-specific hazard rates in \code{onset_hazard} and \code{death_hazard}.}

\item{RR}{Numeric. The relative risk of disease onset for individuals who have inherited the rare variant.}

\item{FamID}{Numeric. The family ID to assign to the simulated pedigree.}

\item{founder_byears}{Numeric list of length 2.  The span of years from which to simulate, uniformly, the founder's birth year.}

\item{stop_year}{Numeric. The last year of study.  If missing the current year is used.}

\item{birth_range}{Numeric list of length 2. By default, \code{birth_range = c(18, 45)}. The minimum and maximum allowable ages that individuals will be able to reproduce.}

\item{NB_params}{Numeric list of length 2. By default, \code{NB_params = c(2, 4/7)}. The size and probability parameters of the negative binomial distribution used to model the number of children per household.}
}
\value{
The ped file for the simulated pedigree.
}
\description{
\code{sim_ped} randomly generates a pedigree.
}
\details{
Please note distinction between \code{sim_ped} and \code{sim_RVpedigree}.  Pedigrees simulated using \code{sim_ped} do not account for study design.  To simulate a pedigree ascertained for a family-based study please use \code{\link{sim_RVpedigree}}.

By assumption, all pedigrees are segregating a rare variant, which is rare enough to have been introduced by exactly one founder.  \code{sim_ped} starts simulating the pedigree by generating the birth year for this founder, uniformly between the years specified by \code{founder_byears}.  Next, all life events are simulated for the founder.  Possible life events include: reproduction, disease onset, and death.  Currently, \code{sim_ped} only allows disease onset to occur once, i.e. no remission after disease onset.  Compuationally, this implies that after an individual has experienced disease onset, their waiting time to death is always simulated using the age-specific mortaility rates for the \emph{affected} population (i.e. the second column specified in \code{death_hazard}).  Life events for individuals who have inherited the rare variant are simulated such that their relative risk of developing disease is \code{RR}, according to a Cox proportional hazards model. For all individuals who have not inherited the rare variant the relative risk of disease onset is 1.  Any life events that occur after \code{stop_year} are censored.

The rare variant is transmitted to any offspring of the founder according to Mendel's laws, and the process of simulating life events is repeated for offspring, recursively, until no additional offspring are produced.
}
\section{See Also}{

\code{\link{sim_RVpedigree}}
}
\examples{
data(AgeSpecific_Hazards)

my_part <- seq(0, 100, by = 1)
my_onset_haz <- AgeSpecific_Hazards[,1]
my_death_haz <- AgeSpecific_Hazards[,c(2,3)]

#Simulate a random pedigree
set.seed(22)
ex_ped <- sim_ped(onset_hazard = my_onset_haz,
                  death_hazard = my_death_haz,
                  part = my_part,
                  RR = 5, FamID = 1,
                  founder_byears = c(1900, 1910),
                  stop_year = 2015)

#Plot pedigree with the kinship2 package
library(kinship2)

#Define pedigree to use kinship2's plot function
RV_status <- ex_ped$DA1 + ex_ped$DA2
Affected  <- ex_ped$affected

ex_pedigree <- pedigree(id = ex_ped$ID,
                        dadid = ex_ped$dad_id,
                        momid = ex_ped$mom_id,
                        sex = (ex_ped$gender + 1),
                        affected = cbind(Affected, RV_status),
                        famid = ex_ped$FamID)
ExamplePed <- ex_pedigree['1']
plot(ExamplePed)
pedigree.legend(ExamplePed, location = "topleft",  radius = 0.25)

}

