% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedSim_Functions.R
\name{sim_ped}
\alias{sim_ped}
\title{Simulate a pedigree.}
\usage{
sim_ped(onset_hazard, death_hazard, part, RR, FamID, founder_byears, stop_year,
  birth_range = c(18, 45), NB_params = c(2, 4/7))
}
\arguments{
\item{onset_hazard}{Numeric vector. The population age-specific onset hazard.}

\item{death_hazard}{data.frame. Column 1 should specify the age specific
mortality rates in the unaffected population, while column 2 should provide
the age specific morality rates in the affected population.}

\item{part}{A numeric vector.  Partition of ages over which to apply the
age-specific hazards}

\item{RR}{A numeric constant. The relative risk of developing
disease for individuals who have inherited the rare variant.}

\item{FamID}{Numeric. Family identification number.}

\item{founder_byears}{A numeric vector of length 2, the minimum and maximum founder birth years.  See details.}

\item{stop_year}{The last year of study.  Note: any information simulated after stop.year is censored and will not be recorded in the ped file.}

\item{birth_range}{A numeric vector of length 2.  By default, \code{birth_range} = c(18, 45).  The minimum and maximum allowable birth ages in simulation.}

\item{NB_params}{A numeric vector of length 2. By default \code{NB_params} = c(2, 4/7). The size and probabiliy parameters of the negative binomial distribution that describes the number of children per household in the population.}
}
\value{
The ped file for the simulated pedigree.
}
\description{
\code{sim_ped} randomly generates a pedigree.
}
\details{
Please note distinction between \code{sim_ped} and \code{sim_RVpedigree}.  Pedigrees simulated using \code{sim_ped} do not account for study design, i.e. no requirement for multiple affected individuals in pedigree.  To simulate a pedigree likely to be ascertained according to a particular study design please use \code{\link{sim_RVpedigree}}.

By assumption, all ascertained pedigrees are segregating a rare variant, which is rare enough to have been introduced by exactly one founder.  \code{sim_RVpedigree} starts the pedigree simulation process by simulating the birth year,  uniformly between the years specified by \code{founder_byears}, for the founder who has introduced the rare variant into the pedigree.  Next, life events are simulated for the founder.  Possible life events include: reproduction, disease onset, and death.  Currently, \code{sim_RVpedigree} only allows disease onset to occur once, i.e. no remission after disease onset.  Compuationally, this implies that after an individual has experienced disease onset, their waiting time to death is always simulated using the age-specific mortaility rates for the \emph{affected} population (i.e. the second column specified in \code{death_hazard}).  Life events for individuals who have inherited the rare variant are simulated such that their relative risk of developing disease is \code{RR}, according to a Cox proportional hazards model. For all individuals who have not inherited the rare variant, we assign a relative risk of disease onset of 1.  Any life events that occur after \code{stop_year} are censored.

The rare variant is transmitted to any offspring of the founder according to Mendel's laws, and the process of simulating life events is repeated for offspring, recursively, until no additional offspring are produced.
}
\section{See Also}{

\code{\link{sim_RVpedigree}}
}
\examples{
#Define a parition of ages over which to apply the age-specific hazards
part_vec <- seq(0, 100, by = 1)

#Specify age-specific mortality rates for affected and unaffected individuals
unaffected_mort <- 0.00001 + pgamma(seq(0.16, 16, by = .16),
                                    shape = 9.5, scale = 1)/350
affected_mort <- c(0.55, 0.48, 0.37, 0.23, 0.15,
                   pgamma(seq(0.96, 16, by = .16), shape = 4, scale = 1.5))/300

#Combine mortality rates in data frame, with
# unaffected mortality rates in column 1, and
# affected mortality rates in column 2

Dhaz_df  <- (as.data.frame(cbind(unaffected_mort, affected_mort)))

#Specify the population age-specific onset hazard
Ohaz_vec <- (dgamma(seq(0.1, 10, by = .1), shape = 8, scale = 0.75))

#Simulate a random pedigree
set.seed(22)
ex_ped <- sim_ped(onset_hazard = Ohaz_vec,
                  death_hazard = Dhaz_df,
                  part = part_vec,
                  RR = 1, FamID = 1,
                  founder_byears = c(1900, 1910),
                  stop_year = 2015)

#Plot pedigree usign kinship2 package
library(kinship2)

#Define pedigree to use kinship2's plot function
RV_status <- ex_ped$DA1 + ex_ped$DA2
Affected  <- ex_ped$affected

ex_pedigree <- pedigree(id = ex_ped$ID,
                        dadid = ex_ped$dad_id,
                        momid = ex_ped$mom_id,
                        sex = (ex_ped$gender + 1),
                        affected = cbind(Affected, RV_status),
                        famid = ex_ped$FamID)
ExamplePed <- ex_pedigree['1']
plot(ExamplePed)
pedigree.legend(ExamplePed, location = "topleft",  radius = 0.25)

}

