% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedSim_Functions.R
\name{sim_ped}
\alias{sim_ped}
\title{Simulate a pedigree}
\usage{
sim_ped(hazard_rates, part, GRR, prob_causalRV, FamID, founder_byears,
  stop_year, single_founderEntry = TRUE, birth_range = c(18, 45),
  NB_params = c(2, 4/7))
}
\arguments{
\item{hazard_rates}{Data.frame. Column 1 should specify the population age-specific hazard rate for disease, column 2 should specify the age-specific hazard rate for death in the unaffected population, and column 3 should specify the age-specific hazard rate for death in the affected population. See details.}

\item{part}{Numeric vector. The partition of ages over which to apply the age-specific hazard rates in \code{hazard_rates}.}

\item{GRR}{Numeric. The relative-risk of disease for individuals who inherit the rare variant.}

\item{FamID}{Numeric. The family ID to assign to the simulated pedigree.}

\item{founder_byears}{Numeric vector of length 2.  The span of years from which to simulate, uniformly, the birth year for the founder who introduced the rare variant to the pedigree.}

\item{stop_year}{Numeric. The last year of study.  If missing, the current year is used.}

\item{single_founderEntry}{Logical. When \code{single_founderEntry = TRUE}, only 1 founder can introduce a causal variant to the pedigee; when \code{single_founderEntry = FALSE} multiple founders may introduce the causal variant.  By default, \code{single_founderEntry = TRUE}.}

\item{birth_range}{Numeric vector of length 2. The minimum and maximum allowable ages, in years, between which individuals may reproduce.  By default, \code{birth_range = c(18, 45)}.}

\item{NB_params}{Numeric vector of length 2. The size and probability parameters of the negative binomial distribution used to model the number of children per household.  By default, \code{NB_params = c(2, 4/7)}, due to the investigation of Kojima and Kelleher (1962).}
}
\value{
The simulated pedigree.
}
\description{
Please note the distinction between \code{sim_ped} and \code{sim_RVped}.  Pedigrees simulated using \code{sim_ped} do not account for study design.  To simulate a pedigree ascertained to contain multiple family members affected by a disease please use \code{\link{sim_RVped}}.
}
\details{
By assumption, all pedigrees are segregating a rare variant, which is rare enough to have been introduced by exactly one founder.  \code{sim_ped} starts simulating the pedigree by generating the birth year for this founder, uniformly between the years specified by \code{founder_byears}.  Next, all life events are simulated for the founder via \code{\link{sim_lifeEvents}}.  Possible life events include: reproduction, disease onset, and death.  We only allow disease onset to occur once, i.e. no remission.  Computationally, this implies that after an individual has experienced disease onset, their waiting time to death is always simulated using the age-specific mortality rates for the \emph{affected} population (i.e. the second column specified in \code{death_hazard}).  Life events for individuals who have inherited the rare variant are simulated such that their relative-risk of developing disease is \code{GRR}, according to a proportional hazards model. For all individuals who have not inherited the rare variant the relative-risk of disease onset is 1.  Any life events that occur after \code{stop_year} are censored.

The rare variant is transmitted to any offspring of the founder according to Mendel's laws, and the process of simulating life events is repeated for offspring, recursively, until no additional offspring are produced.
}
\section{See Also}{

\code{\link{sim_RVped}}
}

\examples{
data(AgeSpecific_Hazards)

my_part <- seq(0, 100, by = 1)

#Simulate a random pedigree
set.seed(22)
ex_ped <- sim_ped(hazard_rates = AgeSpecific_Hazards,
                  part = seq(0, 100, by = 1),
                  GRR = 5, prob_causalRV = 1,
                  FamID = 1,
                  founder_byears = c(1900, 1910),
                  stop_year = 2015)

#Plot pedigree with the kinship2 package
library(kinship2)

#Define pedigree and pass to plot function
RV_status <- ex_ped$DA1 + ex_ped$DA2
Affected  <- ex_ped$affected

ex_pedigree <- pedigree(id = ex_ped$ID,
                        dadid = ex_ped$dadID,
                        momid = ex_ped$momID,
                        sex = (ex_ped$sex + 1),
                        affected = cbind(Affected, RV_status))
plot(ex_pedigree)
pedigree.legend(ex_pedigree, location = "topleft",  radius = 0.25)

}
\references{
OUR MANUSCRIPT

Ken-Ichi Kojima, Therese M. Kelleher. (1962), \emph{Survival of Mutant Genes}. The American Naturalist 96, 329-346.
}
