% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedSim_Functions.R
\name{sim_RVped}
\alias{sim_RVped}
\title{Simulate a Pedigree Ascertained to Contain Multiple Family Members Affected by a Disease}
\usage{
sim_RVped(onset_hazard, death_hazard, part, RR, founder_byears, ascertain_span,
  num_affected, FamID, recall_probs, stop_year, birth_range = c(18, 45),
  NB_params = c(2, 4/7))
}
\arguments{
\item{onset_hazard}{Numeric. The population age-specific hazard rate for disease.}

\item{death_hazard}{Data.frame. Column 1 should specify the age-specific hazard rate for death in the unaffected population, and column 2 should specify the age-specific hazard rate for death in the affected population. See details.}

\item{part}{Numeric. The partition of ages over which to apply the age-specific hazard rates in \code{onset_hazard} and \code{death_hazard}.}

\item{RR}{Numeric. The relative risk of disease for individuals who inherit the rare variant.}

\item{founder_byears}{Numeric list of length 2.  The span of years from which to simulate, uniformly, the birth year for the founder who introduced the rare variant to the pedigree.}

\item{ascertain_span}{Numeric list of length 2.  The year span of the ascertainment period.  This period represents the range of years during which the proband developed disease and the family would have been ascertained for multiple affected relatives.}

\item{num_affected}{Numeric.  The minimum number of affected individuals in the pedigree.}

\item{FamID}{Numeric. The family ID to assign to the simulated pedigree.}

\item{recall_probs}{Numeric. The proband's recall probabilities for relatives, see details.  If missing, four times kinship coefficient between the proband and the relative is used.}

\item{stop_year}{Numeric. The last year of study.  If missing, the current year is used.}

\item{birth_range}{Numeric list of length 2. The minimum and maximum allowable ages, in years, between which individuals may reproduce.  By default, \code{birth_range = c(18, 45)}.}

\item{NB_params}{Numeric list of length 2. The size and probability parameters of the negative binomial distribution used to model the number of children per household.  By default, \code{NB_params = c(2, 4/7)}, due to the investigation of Kojima and Kelleher (1962).}
}
\value{
A list containing the following data frames:

\code{full_ped} The full pedigree, prior to proband selection and trimming.

\code{ascertained_ped} The ascertained pedigree, with proband selected and trimmed according to proband recall probability.
}
\description{
\code{sim_RVped} simulates a pedigree ascertained to contain multiple affected members, selects a proband, and trims the pedigree to contain only those individuals that are recalled by the proband.
}
\details{
By assumption, all simulated pedigrees are segregating a genetic susceptibility variant.  We assume that the variant is rare enough that it has been introduced by one founder.  We begin the simulation of the pedigree with this founder, and transmit the rare variant from parent to offspring according to Mendel's laws.

\code{sim_RVped} begins pedigree simulation by generating the year of birth, uniformly, between the years specified in \code{founder_byears} for the founder who introduced the rare variant to the pedigree.  Next, we simulate this founder's life events using the internal \code{\link{sim_lifeEvents}} function, and censor any events that occur after the study \code{stop_year}.  Possible life events include: reproduction, disease onset, and death. We continue simulating life events for any offspring, censoring events which occur after the study stop year, until the simulation process terminates.

We do not model disease remission. Rather, we impose the restriction that individuals may only experience disease onset once, and remain affected from that point on.  After disease onset occurs the affected hazard rate for death is applied.

\code{sim_RVped} will only return ascertained pedigrees with at least \code{num_affected} affected individuals.  That is, if a simulated pedigree does not contain at least \code{num_affected} affected individuals \code{sim_RVped} will discard that pedigree and simulate another until the condition is met.  We note that even for \code{num_affected} \eqn{= 2}, \code{sim_RVped} can be computationally expensive.  To simulate a pedigree with no proband, and without a minimum number of affected members use instead \code{\link{sim_ped}}.

Upon simulating a pedigree with \code{num_affected} individuals, \code{sim_RVped} chooses a proband from the set of available candidates.  Candidates for proband selection must have the following qualities:
\enumerate{
  \item experienced disease onset between the years specified by \code{ascertain_span},
  \item if less than \code{num_affected} - 1 individuals experienced disease onset prior to the lower bound of \code{ascertain_span}, a proband is chosen from the affected individuals, such that there were at least \code{num_affected} affected individuals when the pedigree was ascertained through the proband.
}

After the proband is selected, the pedigree is trimmed based on the proband's recall probability of his or her relatives.  This option is included to allow researchers to model the possibility that a proband either cannot provide a complete family history or that they explicitly request that certain family members not be contacted.  If \code{recall_probs} is missing, the default values of four times the kinship coefficient, as defined by Thompson (see references), between the proband and his or her relatives are assumed.  This has the effect of retaining all first degree relatives with probability 1, retaining all second degree relatives with probability 0.5, retaining all third degree relatives with probability 0.25, etc.  Alternatively, the user may specify a list of length \eqn{l}, such that the first \eqn{l-1} items represent the respective recall probabilities for relatives of degree \eqn{1, 2, ... , l-1} and the \eqn{l^{th}} item represents the recall probability of a relative of degree \eqn{l} or greater. For example, if \code{recall_probs = c(1, 0.75, 0.5)}, then all first degree relatives (i.e. parents, siblings, and offspring) are retained with probability 1, all second degree relatives (i.e. grandparents, grandchildren, aunts, uncles, nieces and nephews) are retained with probability 0.75, and all other relatives are retained with probability 0.5. To simulate fully ascertained pedigrees, simply specify \code{recall_probs = c(1)}.

In the event that a trimmed pedigree fails the \code{num_affected} condition,  \code{sim_RVped} will discard that pedigree and simulate another until the condition is met.  For this reason, the values specified for \code{recall_probs} affect computation time.
}
\section{See Also}{

\code{\link{sim_ped}}, \code{\link{trim_ped}}, \code{\link{sim_lifeEvents}}
}
\examples{
#Read in age-specific hazards
data(AgeSpecific_Hazards)

#Simulate pedigree ascertained for multiple affected individuals
set.seed(13)
ex_RVped <- sim_RVped(onset_hazard = AgeSpecific_Hazards[,1],
                      death_hazard = AgeSpecific_Hazards[,c(2,3)],
                      part = seq(0, 100, by = 1),
                      RR = 15, FamID = 1,
                      founder_byears = c(1900, 1910),
                      ascertain_span = c(1900, 2015),
                      num_affected = 2, stop_year = 2015,
                      recall_probs = c(1, 0.75, 0.5))




#Plot ex_RVped pedigrees using the kinship2 package
library(kinship2)

#Define pedigree object for full pedigree, i.e. pedigree prior to trimming
FullRVped <- pedigree(id = ex_RVped[[1]]$ID,
                        dadid = ex_RVped[[1]]$dad_id,
                        momid = ex_RVped[[1]]$mom_id,
                        sex = (ex_RVped[[1]]$gender + 1),
                        affected = cbind(Affected = ex_RVped[[1]]$affected,
                                         RV_status = ex_RVped[[1]]$DA1 +
                                                     ex_RVped[[1]]$DA2),
                        famid = ex_RVped[[1]]$FamID)

#plot pedigree and legend
plot(FullRVped['1'])
pedigree.legend(FullRVped, location = "topleft",  radius = 0.25)


#Define pedigree object for trimmed pedigree, i.e, pedigree with
# proband selected and relatives trimmed
TrimRVped <- pedigree(id = ex_RVped[[2]]$ID,
                       dadid = ex_RVped[[2]]$dad_id,
                       momid = ex_RVped[[2]]$mom_id,
                       sex = (ex_RVped[[2]]$gender + 1),
                       affected = cbind(Affected = ex_RVped[[2]]$affected,
                                        Proband = ex_RVped[[2]]$proband,
                                        RV_status = ex_RVped[[2]]$DA1 +
                                                    ex_RVped[[2]]$DA2),
                       famid = ex_RVped[[2]]$FamID)['1']
plot(TrimRVped)
pedigree.legend(TrimRVped, location = "topleft",  radius = 0.25)


}
\references{
OUR MANUSCRIPT

Ken-Ichi Kojima, Therese M. Kelleher. (1962), \emph{Survival of Mutant Genes}. The American Naturalist 96, 329-346.

Thompson, E. (2000). \emph{Statistical Inference from Genetic Data on Pedigrees.} NSF-CBMS Regional Conference Series in Probability and Statistics, 6, I-169. Retrieved from http://www.jstor.org.proxy.lib.sfu.ca/stable/4153187
}

