% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedSim_Functions.R
\name{sim_RVpedigree}
\alias{sim_RVpedigree}
\title{Simulate a pedigree ascertained for a given number of affected.}
\usage{
sim_RVpedigree(onset_hazard, death_hazard, part, RR, founder_byears,
  ascertain_span, num_affected, family_num, recall_probs, birth_range = c(18,
  45), NB_params = c(2, 4/7), stop_year = 2015)
}
\arguments{
\item{onset_hazard}{Numeric vector. The population age-specific onset hazard.}

\item{death_hazard}{data.frame. Column 1 should specify the age specific
mortality rates in the unaffected population, while column 2 should provide
the age specific morality rates in the affected population.}

\item{part}{A numeric vector.  Partition of ages over which to apply the
age-specific hazards}

\item{RR}{A numeric constant. The relative risk of developing
disease for individuals who have inherited the rare variant.}

\item{founder_byears}{A numeric vector of length 2, the minimum and maximum founder birth years.  See details.}

\item{ascertain_span}{numeric vector of length 2. The ascertainment period,
i.e. first and last year of study.}

\item{num_affected}{numeric. The desired number of affected.}

\item{family_num}{numeric. The family number to assign the simulated RV pedigree.}

\item{recall_probs}{numeric.  Proband recall probabilities.  See details.}

\item{birth_range}{A numeric vector of length 2.  By default, \code{birth_range} = c(18, 45).  The minimum and maximum allowable birth ages in simulation.}

\item{NB_params}{A numeric vector of length 2. By default \code{NB_params} = c(2, 4/7). The size and probabiliy parameters of the negative binomial distribution that describes the number of children per household in the population.}

\item{stop_year}{The last year of study.  By default, \code{stop_year} = 2015. Note: any information simulated after stop.year is censored and will not be recorded in the ped file.}
}
\value{
full_ped The full ped file, prior to proband selection and trimming.

trim_ped The trimmed ped file, trimmed according to proband recall probability.
}
\description{
\code{sim_RVpedigree} simulates a pedigree with the desired number of
affected, then chooses a proband from the available candidates, or the
individuals who have become affected during the acertainment period of the
study, and finally trims the pedigree based on recall_probs
}
\examples{
part_vec <- seq(0, 100, by = 1)
unaffected_mort <- 0.00001 + pgamma(seq(0.16, 16, by = .16),
                                    shape = 9.5, scale = 1)/350
affected_mort <- c(0.55, 0.48, 0.37, 0.23, 0.15,
                   pgamma(seq(0.96, 16, by = .16), shape = 4, scale = 1.5))/300
Dhaz_df  <- (as.data.frame(cbind(unaffected_mort, affected_mort)))
Ohaz_vec <- (dgamma(seq(0.1, 10, by = .1), shape = 8, scale = 0.75))
set.seed(22)
ex_RVped <- sim_RVpedigree(onset_hazard = Ohaz_vec, death_hazard = Dhaz_df,
                           part = part_vec, RR = 5,
                           founder_byears = c(1900, 1910),
                           ascertain_span = c(1900, 2015),
                           num_affected = 2, family_num = 1)

plot.RVpedigree(ex_RVped[[1]])
plot.RVpedigree(ex_RVped[[2]])

}

