% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedSim_Functions.R
\name{sim_RVpedigree}
\alias{sim_RVpedigree}
\title{Simulate a pedigree ascertained for a given number of affected.}
\usage{
sim_RVpedigree(onset_hazard, death_hazard, part, RR, founder_byears,
  ascertain_span, num_affected, family_num, recall_probs, birth_range = c(18,
  45), NB_params = c(2, 4/7), stop_year = 2015)
}
\arguments{
\item{onset_hazard}{Numeric vector. The population age-specific onset hazard.}

\item{death_hazard}{data.frame. Column 1 should specify the age specific
mortality rates in the unaffected population, while column 2 should provide
the age specific morality rates in the affected population.}

\item{part}{A numeric vector.  Partition of ages over which to apply the
age-specific hazards}

\item{RR}{A numeric constant. The relative risk of developing
disease for individuals who have inherited the rare variant.}

\item{founder_byears}{A numeric vector of length 2, the minimum and maximum founder birth years.  See details.}

\item{ascertain_span}{numeric vector of length 2. The ascertainment period,
i.e. first and last year of study.}

\item{num_affected}{Numeric.  The desired number of affected.}

\item{family_num}{numeric. The family number to assign the simulated RV pedigree.}

\item{recall_probs}{Numeric.  Proband recall probabilities.  See details.}

\item{birth_range}{A numeric vector of length 2.  By default, \code{birth_range} = c(18, 45).  The minimum and maximum allowable birth ages in simulation.}

\item{NB_params}{A numeric vector of length 2. By default \code{NB_params} = c(2, 4/7). The size and probabiliy parameters of the negative binomial distribution that describes the number of children per household in the population.}

\item{stop_year}{The last year of study.  By default, \code{stop_year} = 2015. Note: any information simulated after stop.year is censored and will not be recorded in the ped file.}
}
\value{
full_ped The full ped file, prior to proband selection and trimming.

trim_ped The trimmed ped file, trimmed according to proband recall probability.
}
\description{
\code{sim_RVpedigree} simulates a pedigree with the desired number of
affected, chooses a proband,and trims the pedigree based on the proband's
recall probability of his or her relatives
}
\details{
We assume that the variant is rare enough that it has only been introduced to the pedigree by one of the two oldest founders.  The variant is then transmitted to offspring according to Mendel's  laws.  Life events for individuals who have inherited the rare variant (RV) are simulated such that their relative risk is \code{RR}, all others are simulated with a relative risk of 1.  All events which occur after the specified \code{stop_year} are censored.

By default \code{recall_probs} is 4 times the kinship coefficent between the proband and the probands relative, which results in a recall probability of \eqn{2^{-(n-1)}} for a relative of degree \eqn{n}. Alternatively, the user may specify a list of recall probabilites of length \eqn{l > 0}, in which case the first \emph{l-1} items in \code{recall_probs} are the respective proband recall probabilites for relatives of degree \emph{1, 2, ..., l-1}, and the \emph{l}th item in \code{recall_probs} is the proband's recall probability for all relatives of degree \strong{\emph{l} or greater}.  For example if \code{recall_probs = c(1)} all relatives will be recalled by the proband with probability 1.

The birth year for the founder with the rare variant is simulated uniformly  between the years specified in \code{founder_byears}.
}
\examples{
part_vec <- seq(0, 100, by = 1)
unaffected_mort <- 0.00001 + pgamma(seq(0.16, 16, by = .16),
                                    shape = 9.5, scale = 1)/350
affected_mort <- c(0.55, 0.48, 0.37, 0.23, 0.15,
                   pgamma(seq(0.96, 16, by = .16), shape = 4, scale = 1.5))/300
Dhaz_df  <- (as.data.frame(cbind(unaffected_mort, affected_mort)))
Ohaz_vec <- (dgamma(seq(0.1, 10, by = .1), shape = 8, scale = 0.75))
set.seed(22)
ex_RVped <- sim_RVpedigree(onset_hazard = Ohaz_vec, death_hazard = Dhaz_df,
                           part = part_vec, RR = 5,
                           founder_byears = c(1900, 1910),
                           ascertain_span = c(1900, 2015),
                           num_affected = 2, family_num = 1)

plot_RVpedigree(ex_RVped[[1]])
plot_RVpedigree(ex_RVped[[2]])

}

