% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimRVPedigree_doc.R
\docType{package}
\name{SimRVPedigree}
\alias{SimRVPedigree}
\alias{SimRVPedigree-package}
\title{Randomly Simulate Pedigrees Ascertained for Disease Status.}
\description{
The \code{SimRVPedigree} package allows random simulation of pedigrees likely to be asertained for multiple affected members under a particular study design.  Pedigrees include individual specific variables such as birth year, and, when applicable, onset year, and/or death year.  Additionally, we simualte the transmission of a rare variant, assumed to increase disease susceptibility, according to Mendel's laws.
}
\details{
In addition to randomly simulating pedigrees, \code{SimRVPedigree} also includes functions for: simulating waiting times associated with a non-homogeneous Poisson process, randomly simulating life events contingent on user-provided age-specific hazards.
}

