% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimRVPedigree_doc.R
\docType{package}
\name{SimRVPedigree}
\alias{SimRVPedigree}
\alias{SimRVPedigree-package}
\title{Randomly Simulate Pedigrees Ascertained for Disease Status.}
\description{
The \code{SimRVPedigree} package provides methods to randomly simulate and manipulate pedigrees ascertained to contain multiple family members affected by a rare disease.
}
\details{
Family-based studies designed to identify genetic susceptibility factors associated with rare diseases have received significant consideration in recent years.  Due to the rarity of the diseases under consideration, in some cases, garnering a suitable number of families for analysis could require decades of continued collaboration between researchers and clinicians.  Understandably, due to the nature of these studies, they are inherently difficult and sometimes infeasible to replicate.  The \code{SimRVPedigree} package aims to address this problem by providing a platform to randomly simulate pedigrees similar to those ascertained  for family-based studies to contain multiple relatives affected by a rare disease.

The distinguishing feature of the \code{SimRVPedigree} package is that it aims to accurately mimic the process of family development, while allowing users to incorporate multiple facets of a particular ascertainment scheme.
}

