% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trimming_Functions.R
\name{trim_ped}
\alias{trim_ped}
\title{Trim Pedigree Based on Proband Recall.}
\usage{
trim_ped(ped_file, recall_probs)
}
\arguments{
\item{ped_file}{Data.frame. A pedigree generated by \code{sim_ped}.}

\item{recall_probs}{Numeric. The proband's recall probabilities for relatives, see details.  If missing, four times kinship coefficient between the proband and the relative is used.}
}
\value{
\code{ped_trim} The trimmed pedigree.
}
\description{
Primarily intended as an internal function, \code{trim_ped} chooses a proband and trims relatives based on the proband's probability of recalling his or her relatives.
}
\details{
By default \code{recall_probs} is four times the kinship coefficient, as defined by Thompson (see references), between the proband and the probands relative, which results in a recall probability of \eqn{2^{-(n-1)}} for a relative of degree \eqn{n}. Alternatively, the user may specify a list of recall probabilities of length \eqn{l > 0}, in which case the first \emph{l-1} items in \code{recall_probs} are the respective proband recall probabilities for relatives of degree \emph{1, 2, ..., l-1}, and the \emph{l}th item in \code{recall_probs} is the proband's recall probability for all relatives of degree \strong{\emph{l} or greater}.  For example if \code{recall_probs = c(1)} all relatives will be recalled by the proband with probability 1.

Occasionally, a trimmed family member must be retained to ensure that the pedigree can be plotted.  When this occurs, family members who share a non-zero kinship coefficient with the proband are censored of all pertinent information, and will always have the following qualities:
\enumerate{
  \item availability status = 0
  \item affected status = NA
  \item birth year = NA
  \item onset year = NA
  \item death year = NA
  \item RR = NA
  \item DA1 = NA
  \item DA2 = NA
}
}
\examples{
#Read in example pedigree to trim
data(EgPeds)

#plot example_ped using kinship2
library(kinship2)
ex_pedigree <- pedigree(id = EgPeds$ID,
                        dadid = EgPeds$dad_id,
                        momid = EgPeds$mom_id,
                        sex = (EgPeds$gender + 1),
                        affected = cbind(Affected = EgPeds$affected,
                                         Proband = EgPeds$proband,
                                         RV_status = EgPeds$DA1 +
                                                     EgPeds$DA2),
                        famid = EgPeds$FamID)['1']
plot(ex_pedigree)
pedigree.legend(ex_pedigree, location = "topleft",  radius = 0.25)
mtext("Original Pedigree", side = 3, line = 2)


## Trim pedigree examples
#illustrate the effects of various recall_probs settings
Recall_Probabilities <- list(c(1),
                             c(1, 0),
                             c(1, 0.75, 0.5),
                             c(0.5, 0.1, 0.05))


for (k in 1:length(Recall_Probabilities)) {
   set.seed(2)
   #trim pedigree
   TrimPed <- trim_ped(ped_file = EgPeds[which(EgPeds$FamID == 1), ],
                       recall_probs = Recall_Probabilities[[k]])

   #plot trimmed pedigree
   Tped <- pedigree(id = TrimPed$ID,
                    dadid = TrimPed$dad_id,
                    momid = TrimPed$mom_id,
                    sex = (TrimPed$gender + 1),
                    affected = cbind(Affected = TrimPed$affected,
                                     Proband = TrimPed$proband,
                                     RV_status = TrimPed$DA1 + TrimPed$DA2),
                    famid = TrimPed$FamID)['1']

   plot(Tped)
   pedigree.legend(Tped, location = "topleft",  radius = 0.25)
   mtext(paste0("recall_probs = (", sep = "",
                paste(Recall_Probabilities[[k]], collapse = ", "), ')' ),
                side = 3, line = 2 )
}


}
\references{
OUR MANUSCRIPT

Thompson, E. (2000). \emph{Statistical Inference from Genetic Data on Pedigrees.} NSF-CBMS Regional Conference Series in Probability and Statistics, 6, I-169.
}

