% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trimming_Functions.R
\name{trim_pedigree}
\alias{trim_pedigree}
\title{Choose proband and trim pedigree}
\usage{
trim_pedigree(ped_file, ascertain_span, num_affected, recall_probs)
}
\arguments{
\item{ped_file}{A data.frame. The ped file generated by \code{ped_step}.}

\item{ascertain_span}{numeric vector of length 2. The ascertainment period,
i.e. first and last year of study.}

\item{num_affected}{Numeric.  The desired number of affected.}

\item{recall_probs}{Numeric.  Proband recall probabilities.  See details.}
}
\value{
The trimmed ped file
}
\description{
\code{trim_pedigree} chooses a proband from the pedigree and trims relatives
based on the proband's recall probability of that relative.
}
\details{
By default \code{recall_probs} is 4 times the kinship coefficent between the
proband and the probands relative, which results in a recall probability of
\eqn{2^{-(n-1)}} for a relative of degree \eqn{n}. Alternatively, the user may
specify a list of recall probabilites of length \eqn{l > 0}, in which case
the first \emph{l-1} items in \code{recall_probs} are the respective proband
recall probabilites for relatives of degree \emph{1, 2, ..., l-1}, and the
\emph{l}th item in \code{recall_probs} is the proband's recall probability for
all relatives of degree \strong{\emph{l} or greater}.  For example if
\code{recall_probs = c(1)} all relatives will be recalled by the proband with
probability 1.
}
\examples{
part_vec <- seq(0, 100, by = 1)
unaffected_mort <- 0.00001 + pgamma(seq(0.16, 16, by = .16),
                                    shape = 9.5, scale = 1)/350
affected_mort <- c(0.55, 0.48, 0.37, 0.23, 0.15,
                   pgamma(seq(0.96, 16, by = .16), shape = 4, scale = 1.5))/300
Dhaz_df  <- (as.data.frame(cbind(unaffected_mort, affected_mort)))
Ohaz_vec <- (dgamma(seq(0.1, 10, by = .1), shape = 8, scale = 0.75))/50
set.seed(22)
ex_ped <- ped_step(onset_hazard = Ohaz_vec, death_hazard = Dhaz_df,
                   part = part_vec, RR = 5, founder_byears = c(1900, 1910))
ex_ped
trim_pedigree(ped_file = ex_ped, ascertain_span = c(1990, 2015), num_affected = 3)
trim_pedigree(ped_file = ex_ped, ascertain_span = c(1990, 2015),
          num_affected = 3, recall_probs = c(1, 0.9, 0))

}

